% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{AIC.fixest}
\alias{AIC.fixest}
\title{Aikake's an information criterion}
\usage{
\method{AIC}{fixest}(object, ..., k = 2)
}
\arguments{
\item{object}{A `fixest` object. Obtained using the functions [`femlm`], [`feols`] or [`feglm`].}

\item{...}{Optionally, more fitted objects.}

\item{k}{A numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC (i.e. `AIC=-2*LL+k*nparams`).}
}
\value{
It return a numeric vector, with length the same as the number of objects taken as arguments.
}
\description{
This function computes the AIC (Aikake's, an information criterion) from a `fixest` estimation.
}
\details{
The AIC is computed as:
\deqn{AIC = -2\times LogLikelihood + k\times nbParams}
with k the penalty parameter.

You can have more information on this criterion on [`AIC`][stats::AIC].
}
\examples{

# two fitted models with different expl. variables:
res1 = femlm(Sepal.Length ~ Sepal.Width + Petal.Length +
             Petal.Width | Species, iris)
res2 = femlm(Sepal.Length ~ Petal.Width | Species, iris)

AIC(res1, res2)
BIC(res1, res2)

}
\seealso{
See also the main estimation functions [`femlm`], [`feols`] or [`feglm`]. Other statictics methods: [`BIC.fixest`], [`logLik.fixest`], [`nobs.fixest`].
}
\author{
Laurent Berge
}
