% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{as.list.fixest2_multi}
\alias{as.list.fixest2_multi}
\title{Transforms a fixest2_multi object into a list}
\usage{
\method{as.list}{fixest2_multi}(x, ...)
}
\arguments{
\item{x}{A \code{fixest2_multi} object, obtained from a \code{fixest2} estimation leading to
multiple results.}

\item{...}{Not currently used.}
}
\value{
Returns a list containing all the results of the multiple estimations.
}
\description{
Extracts the results from a \code{fixest2_multi} object and place them into a list.
}
\examples{

base <- iris
names(base) <- c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res <- feols(y ~ csw(x1, x2, x3), base, split = ~species)

# All the results at once
as.list(res)

}
\seealso{
The main fixest2 estimation functions: \code{\link{feols}}, \code{\link[=feglm]{fepois}},
\code{\link[=femlm]{fenegbin}}, \code{\link{feglm}}, \code{\link{feNmlm}}. Tools for mutliple fixest2
estimations: \code{\link{summary.fixest2_multi}}, \code{\link{print.fixest2_multi}}, \code{\link{as.list.fixest2_multi}},
\code{\link[fixest2]{sub-sub-.fixest2_multi}}, \code{\link[fixest2]{sub-.fixest2_multi}}.
}
