% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\docType{data}
\name{base_did}
\alias{base_did}
\title{Sample data for difference in difference}
\format{
`base_did` is a data frame with 1,040 observations and 6 variables named `y`, `x1`, `id`, `period`, `post` and `treat`.

\itemize{
\item{y: The dependent variable affected by the treatment.}
\item{x1: An explanatory variable.}
\item{id: Identifier of the individual.}
\item{period: From 1 to 10}
\item{post: Indicator taking value 1 if the period is strictly greater than 5, 0 otherwise.}
\item{treat: Indicator taking value 1 if the individual is treated, 0 otherwise.}

}
}
\source{
This data has been generated from \pkg{R}.
}
\usage{
data(base_did)
}
\description{
This data has been generated to illustrate the use of difference in difference functions in package \pkg{fixest}. This is a balanced panel of 104 individuals and 10 periods. About half the individuals are treated, the treatment having a positive effect on the dependent variable `y` after the 5th period. The effect of the treatment on `y` is gradual.
}
\keyword{datasets}
