% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{BIC.fixest}
\alias{BIC.fixest}
\title{Bayesian information criterion}
\usage{
\method{BIC}{fixest}(object, ...)
}
\arguments{
\item{object}{A `fixest` object. Obtained using the functions [`femlm`], [`feols`] or [`feglm`].}

\item{...}{Optionally, more fitted objects.}
}
\value{
It return a numeric vector, with length the same as the number of objects taken as arguments.
}
\description{
This function computes the BIC (Bayesian information criterion) from a `fixest` estimation.
}
\details{
The BIC is computed as follows:
\deqn{BIC = -2\times LogLikelihood + \log(nobs)\times nbParams}
with k the penalty parameter.

You can have more information on this criterion on [`AIC`][stats::AIC].
}
\examples{

# two fitted models with different expl. variables:
res1 = femlm(Sepal.Length ~ Sepal.Width + Petal.Length +
            Petal.Width | Species, iris)
res2 = femlm(Sepal.Length ~ Petal.Width | Species, iris)

AIC(res1, res2)
BIC(res1, res2)

}
\seealso{
See also the main estimation functions [`femlm`], [`feols`] or [`feglm`]. Other statistics functions: [`AIC.fixest`], [`logLik.fixest`].
}
\author{
Laurent Berge
}
