% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BIC.fixest2}
\alias{BIC.fixest2}
\title{Bayesian information criterion}
\usage{
\method{BIC}{fixest2}(object, ...)
}
\arguments{
\item{object}{A \code{fixest2} object. Obtained using the functions \code{\link{femlm}}, \code{\link{feols}} or \code{\link{feglm}}.}

\item{...}{Optionally, more fitted objects.}
}
\value{
It return a numeric vector, with length the same as the number of objects taken as arguments.
}
\description{
This function computes the BIC (Bayesian information criterion) from a \code{fixest2} estimation.
}
\details{
The BIC is computed as follows:
\deqn{BIC = -2\times LogLikelihood + \log(nobs)\times nbParams}
with k the penalty parameter.

You can have more information on this criterion on \code{\link[stats:AIC]{AIC}}.
}
\examples{

# two fitted models with different expl. variables:
res1 <- femlm(Sepal.Length ~ Sepal.Width + Petal.Length +
  Petal.Width | Species, iris)
res2 <- femlm(Sepal.Length ~ Petal.Width | Species, iris)

AIC(res1, res2)
BIC(res1, res2)

}
\seealso{
See also the main estimation functions \code{\link{femlm}}, \code{\link{feols}} or \code{\link{feglm}}. Other statistics functions: \code{\link{AIC.fixest2}}, \code{\link{logLik.fixest2}}.
}
\author{
Laurent Berge
}
