% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{bread.fixest}
\alias{bread.fixest}
\title{Extracts the bread matrix from fixest objects}
\usage{
\method{bread}{fixest}(x, ...)
}
\arguments{
\item{x}{A \code{fixest} object, obtained for instance from \code{\link[fixest]{feols}}.}

\item{...}{Not currently used.}
}
\value{
Returns a matrix of the same dimension as the number of variables used in the estimation.
}
\description{
Extracts the bread matrix from fixest objects to be used to compute sandwich variance-covariance matrices.
}
\examples{

est = feols(Petal.Length ~ Petal.Width + Sepal.Width, iris)
bread(est)

}
