% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{$.fixest_multi}
\alias{$.fixest_multi}
\title{Extracts the root of a fixest_multi object}
\usage{
\method{$}{fixest_multi}(x, name)
}
\arguments{
\item{x}{A \code{fixest_multi} object, obtained from a \code{fixest} estimation leading to multiple results.}

\item{name}{The name of the root element to select.}
}
\value{
It either returns a \code{fixest_multi} object or a \code{fixest} object it there is only one estimation associated to the root element.
}
\description{
Extracts an element at the root of a fixest_multi object.
}
\examples{

base <- iris
names(base) <- c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res <- feols(y ~ csw(x1, x2, x3), base, split = ~species)

# Let's the results for the setosa species
res$setosa

# now for versicolor
etable(res$versicolor)

}
\seealso{
The main fixest estimation functions: \code{\link[fixest2]{feols}},
\code{\link[fixest2:feglm]{fepois}}, \code{\link[fixest2:femlm]{fenegbin}},
\code{\link[fixest2]{feglm}}, \code{\link[fixest2]{feNmlm}}. Tools for multiple
fixest estimations: \code{\link[fixest2]{summary.fixest_multi}},
\code{\link[fixest2]{print.fixest_multi}},
\code{\link[fixest2]{as.list.fixest_multi}},
\code{\link[fixest2]{sub-sub-.fixest_multi}},
\code{\link[fixest2]{sub-.fixest_multi}},
\code{\link[fixest2]{cash-.fixest_multi}}.
}
