% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R, R/iplot.R
\name{coefplot}
\alias{coefplot}
\alias{iplot}
\title{Plots confidence intervals and point estimates}
\usage{
coefplot(
  ...,
  objects = NULL,
  style = NULL,
  se,
  ci_low,
  ci_high,
  df.t = NULL,
  vcov = NULL,
  cluster = NULL,
  x,
  x.shift = 0,
  horiz = FALSE,
  dict = NULL,
  keep,
  drop,
  order,
  ci.width = "1\%",
  ci_level = 0.95,
  add = FALSE,
  plot_prms = list(),
  pch = c(20, 17, 15, 21, 24, 22),
  col = 1:8,
  cex = 1,
  lty = 1,
  lwd = 1,
  ylim = NULL,
  xlim = NULL,
  pt.pch = pch,
  pt.bg = NULL,
  pt.cex = cex,
  pt.col = col,
  ci.col = col,
  pt.lwd = lwd,
  ci.lwd = lwd,
  ci.lty = lty,
  grid = TRUE,
  grid.par = list(lty = 3, col = "gray"),
  zero = TRUE,
  zero.par = list(col = "black", lwd = 1),
  pt.join = FALSE,
  pt.join.par = list(col = pt.col, lwd = lwd),
  ci.join = FALSE,
  ci.join.par = list(lwd = lwd, col = col, lty = 2),
  ci.fill = FALSE,
  ci.fill.par = list(col = "lightgray", alpha = 0.5),
  ref = "auto",
  ref.line = "auto",
  ref.line.par = list(col = "black", lty = 2),
  lab.cex,
  lab.min.cex = 0.85,
  lab.max.mar = 0.25,
  lab.fit = "auto",
  xlim.add,
  ylim.add,
  only.params = FALSE,
  sep,
  as.multiple = FALSE,
  bg,
  group = "auto",
  group.par = list(lwd = 2, line = 3, tcl = 0.75),
  main = "Effect on __depvar__",
  value.lab = "Estimate and __ci__ Conf. Int.",
  ylab = NULL,
  xlab = NULL,
  sub = NULL,
  i.select = NULL,
  do_iplot = NULL
)

iplot(
  ...,
  i.select = 1,
  objects = NULL,
  style = NULL,
  se,
  ci_low,
  ci_high,
  df.t = NULL,
  vcov = NULL,
  cluster = NULL,
  x,
  x.shift = 0,
  horiz = FALSE,
  dict = NULL,
  keep,
  drop,
  order,
  ci.width = "1\%",
  ci_level = 0.95,
  add = FALSE,
  plot_prms = list(),
  pch = c(20, 17, 15, 21, 24, 22),
  col = 1:8,
  cex = 1,
  lty = 1,
  lwd = 1,
  ylim = NULL,
  xlim = NULL,
  pt.pch = pch,
  pt.bg = NULL,
  pt.cex = cex,
  pt.col = col,
  ci.col = col,
  pt.lwd = lwd,
  ci.lwd = lwd,
  ci.lty = lty,
  grid = TRUE,
  grid.par = list(lty = 3, col = "gray"),
  zero = TRUE,
  zero.par = list(col = "black", lwd = 1),
  pt.join = FALSE,
  pt.join.par = list(col = pt.col, lwd = lwd),
  ci.join = FALSE,
  ci.join.par = list(lwd = lwd, col = col, lty = 2),
  ci.fill = FALSE,
  ci.fill.par = list(col = "lightgray", alpha = 0.5),
  ref = "auto",
  ref.line = "auto",
  ref.line.par = list(col = "black", lty = 2),
  lab.cex,
  lab.min.cex = 0.85,
  lab.max.mar = 0.25,
  lab.fit = "auto",
  xlim.add,
  ylim.add,
  only.params = FALSE,
  sep,
  as.multiple = FALSE,
  bg,
  group = "auto",
  group.par = list(lwd = 2, line = 3, tcl = 0.75),
  main = "Effect on __depvar__",
  value.lab = "Estimate and __ci__ Conf. Int.",
  ylab = NULL,
  xlab = NULL,
  sub = NULL
)
}
\arguments{
\item{...}{Other arguments to be passed to \code{summary}, if \code{object} is an estimation,
and/or to the function \code{plot} or \code{lines} (if \code{add = TRUE}).}

\item{objects}{A list of \code{fixest} estimation objects, or \code{NULL} (default). If provided,
the objects in \code{...} are ignored and the only coefficients reported are the ones in the
argument \code{objects}.}

\item{style}{A character scalar giving the style of the plot to be used. You
can set styles with the function \code{\link{setFixest_coefplot}}, setting all the default
values of the function. If missing, then it switches to either "default" or "iplot",
depending on the calling function.}

\item{se}{The standard errors of the estimates. It may be missing.}

\item{ci_low}{If \code{se} is not provided, the lower bound of the confidence interval.
For each estimate.}

\item{ci_high}{If \code{se} is not provided, the upper bound of the confidence interval.
For each estimate.}

\item{df.t}{Integer scalar or \code{NULL} (default). The degrees of freedom (DoF) to use
when computing the confidence intervals with the Student t. By default it
tries to capture the DoF from the estimation. To use a Normal law to compute the
confidence interval, use \code{df.t = Inf}.}

\item{vcov}{Versatile argument to specify the VCOV.
In general, it is either a character scalar equal to a VCOV type, either a formula of the form:
vcov_type ~ variables. The VCOV types implemented are: "iid", "hetero" (or "HC1"),
"cluster", "twoway", "NW" (or "newey_west"), "DK" (or "driscoll_kraay"), and "conley".
It also accepts object from vcov_cluster, vcov_NW, NW, vcov_DK, DK, vcov_conley and conley.
It also accepts covariance matrices computed externally.
Finally it accepts functions to compute the covariances.
See the vcov documentation in the vignette.

You can pass several VCOVs (as above) if you nest them into a list.
If the number of VCOVs equals the number of models, each VCOV is mapped to the appropriate model.
If there is one model and several VCOVs, or if the first element of the list is equal to
\code{"each"} or \code{"times"}, then the estimations will be replicated and the results
for each estimation and each VCOV will be reported.}

\item{cluster}{Tells how to cluster the standard-errors (if clustering is requested).
Can be either a list of vectors, a character vector of variable names, a formula or
an integer vector. Assume we want to perform 2-way clustering over \code{var1} and \code{var2}
contained in the data.frame \code{base} used for the estimation. All the following
\code{cluster} arguments are valid and do the same thing:
\code{cluster = base[, c("var1", "var2")]}, \code{cluster = c("var1", "var2")}, \code{cluster = ~var1+var2}.
If the two variables were used as fixed-effects in the estimation, you can leave it
blank with \code{vcov = "twoway"} (assuming \code{var1} [resp. \code{var2}] was
the 1st [resp. 2nd] fixed-effect). You can interact two variables using \code{^} with
the following syntax: \code{cluster = ~var1^var2} or \code{cluster = "var1^var2"}.}

\item{x}{The value of the x-axis. If missing, the names of the argument \code{estimate}
are used.}

\item{x.shift}{Shifts the confidence intervals bars to the left or right, depending
on the value of \code{x.shift}. Default is 0.}

\item{horiz}{A logical scalar, default is \code{FALSE}. Whether to display the confidence
intervals horizontally instead of vertically.}

\item{dict}{A named character vector or a logical scalar. It changes the original variable names
to the ones contained in the \code{dict}ionary. E.g. to change the variables named \code{a} and \code{b3} to
(resp.) \dQuote{$log(a)$} and to \dQuote{$bonus^3$}, use \code{dict=c(a="$log(a)$",b3="$bonus^3$")}.
By default, it is equal to \code{getFixest_dict()}, a default dictionary which can be set with
\code{\link{setFixest_dict}}. You can use \code{dict = FALSE} to disable it. By default \code{dict} modifies the
entries in the global dictionary, to disable this behavior, use "reset" as the first element
(ex: \code{dict=c("reset", mpg="Miles per gallon")}).}

\item{keep}{Character vector. This element is used to display only a subset of variables. This
should be a vector of regular expressions (see \code{\link[base:regex]{base::regex}} help for more info). Each
variable satisfying any of the regular expressions will be kept. This argument is applied post
aliasing (see argument \code{dict}).
Use the argument \code{keep_raw} for the same effect before aliasing.

Example: you have the variable \code{x1} to \code{x55} and want to display
only \code{x1} to \code{x9}, then you could use \code{keep = "x[[:digit:]]$"}. If the first character is an
exclamation mark, the effect is reversed (e.g. keep = "!Constant" means: every variable that
does not contain \dQuote{Constant} is kept). See details.}

\item{drop}{Character vector. This element is used if some variables are not to be displayed.
This should be a vector of regular expressions (see \code{\link[base:regex]{base::regex}} help for more info). Each
variable satisfying any of the regular expressions will be discarded. This argument is applied
post aliasing (see argument \code{dict}).
Use the argument \code{drop_raw} for the same effect before aliasing.

Example: you have the variable \code{x1} to \code{x55} and want to
display only \code{x1} to \code{x9}, then you could use \verb{drop = "x[[:digit:]]\{2\}}". If the first character
is an exclamation mark, the effect is reversed (e.g. drop = "!Constant" means: every variable
that does not contain \dQuote{Constant} is dropped). See details.}

\item{order}{Character vector. This element is used if the user wants the variables to be
ordered in a certain way. This should be a vector of regular expressions (see \code{\link[base:regex]{base::regex}}
help for more info). The variables satisfying the first regular expression will be placed first,
then the order follows the sequence of regular expressions. This argument is applied post
aliasing (see argument \code{dict}). Use the argument \code{order_raw} for the same effect before aliasing.

Example: you have the following variables: \code{month1} to \code{month6},
then \code{x1} to \code{x5}, then \code{year1} to \code{year6}. If you want to display first the x's, then the
years, then the months you could use: \code{order = c("x", "year")}. If the first character is an
exclamation mark, the effect is reversed (e.g. order = "!Constant" means: every variable that
does not contain \dQuote{Constant} goes first).  See details.}

\item{ci.width}{The width of the extremities of the confidence intervals. Default is \code{0.1}.}

\item{ci_level}{Scalar between 0 and 1: the level of the CI. By default it is equal to 0.95.}

\item{add}{Default is \code{FALSE}, if the intervals are to be added to an existing
graph. Note that if it is the case, then the argument \code{x} MUST be numeric.}

\item{plot_prms}{A named list. It may contain additionnal parameters to be passed
to the plot.}

\item{pch}{The pch of the coefficient estimates. Default is 1 (circle).
This is an alias to the argument \code{pt.pch}.}

\item{col}{The color of the points and the confidence intervals. Default is 1
("black"). Note that you can set the colors separately for each of them
with \code{pt.col} and \code{ci.col}.}

\item{cex}{Numeric, default is 1. Expansion factor for the points}

\item{lty}{The line type of the confidence intervals. Default is 1.
This is an alias to the argument \code{ci.lty}.}

\item{lwd}{General line width. Default is 1.}

\item{ylim}{Numeric vector of length 2 which gives the limits of the plotting region for
the y-axis. The default is \code{NULL}, which means that it is automatically defined.
Use the argument \code{ylim.add} to simply increase or decrease the default limits.}

\item{xlim}{Numeric vector of length 2 which gives the limits of the plotting region for
the x-axis. The default is \code{NULL}, which means that it is automatically defined.
Use the argument \code{xlim.add} to simply increase or decrease the default limits.}

\item{pt.pch}{The pch of the coefficient estimates. Default is 1 (circle).}

\item{pt.bg}{The background color of the point estimate (when the \code{pt.pch} is
in 21 to 25). Defaults to NULL.}

\item{pt.cex}{The size of the coefficient estimates. Default is the other argument \code{cex}.}

\item{pt.col}{The color of the coefficient estimates. Default is equal to the argument \code{col}.}

\item{ci.col}{The color of the confidence intervals. Default is equal to the argument \code{col}.}

\item{pt.lwd}{The line width of the coefficient estimates. Default is equal to
the other argument \code{lwd}.}

\item{ci.lwd}{The line width of the confidence intervals. Default is equal to
the other argument \code{lwd}.}

\item{ci.lty}{The line type of the confidence intervals. Default is 1.}

\item{grid}{Logical, default is \code{TRUE}. Whether a grid should be displayed. You
can set the display of the grid with the argument \code{grid.par}.}

\item{grid.par}{List. Parameters of the grid. The default values are: \code{lty =  3} and \code{col = "gray"}. You can add any graphical parameter that will be passed
to \code{\link[graphics:abline]{graphics::abline}}. You also have two additional arguments: use \code{horiz =  FALSE} to disable the horizontal lines, and use \code{vert = FALSE} to disable the
vertical lines. Eg: \code{grid.par = list(vert = FALSE, col = "red", lwd = 2)}.}

\item{zero}{Logical scalar, default is \code{TRUE}. Whether the 0 should be displayed
in the limits of the y-axis.
Note that you can set how this zero line looks like with the argument \code{zero.par}.}

\item{zero.par}{A named list of graphical parameters or a logical scalar.
This argument is a list containing the graphical parameters used to draw the zero-line.
The default value is \code{list(col = "black", lwd = 1)} (it's the same if \code{TRUE}).
Set it to \code{FALSE} to turn off the special emphasis of the zero line.
You can add any graphical parameter that will be passed
to \code{\link[graphics:abline]{graphics::abline}}. Example: \code{zero.par = list(col = "darkblue", lwd = 3)}.}

\item{pt.join}{Logical, default is \code{FALSE}. If \code{TRUE}, then the coefficient estimates
are joined with a line.}

\item{pt.join.par}{List. Parameters of the line joining the coefficients. The
default values are: \code{col = pt.col} and \code{lwd = lwd}. You can add any graphical
parameter that will be passed to \code{\link{lines}}. Eg: \code{pt.join.par = list(lty = 2)}.}

\item{ci.join}{Logical default to \code{FALSE}. Whether to join the extremities of
the confidence intervals. If \code{TRUE}, then you can set the graphical parameters
with the argument \code{ci.join.par}.}

\item{ci.join.par}{A list of parameters to be passed to \code{\link[graphics:lines]{graphics::lines}}.
Only used if \code{ci.join=TRUE}. By default it is equal to \code{list(lwd = lwd, col = col, lty = 2)}.}

\item{ci.fill}{Logical default to \code{FALSE}. Whether to fill the confidence intervals
with a color. If \code{TRUE}, then you can set the graphical parameters
with the argument \code{ci.fill.par}.}

\item{ci.fill.par}{A list of parameters to be passed to \code{\link[graphics:polygon]{graphics::polygon}}.
Only used if \code{ci.fill=TRUE}. By default it is equal to \code{list(col = "lightgray", alpha = 0.5)}.
Note that \code{alpha} is a special parameter that adds transparency to the color
(ranges from 0 to 1).}

\item{ref}{Used to add points at \code{y = 0} (typically to visualize reference points).
Either: i) "auto" (default), ii) a character vector of length 1, iii) a list
of length 1, iv) a named integer vector of length 1, or v) a numeric vector.
By default, in \code{iplot}, if the argument \code{ref} has been used in the estimation,
these references are automatically added. If ii), ie a character scalar, then
that coefficient equal to zero is added as the first coefficient. If a list or
a named integer vector of length 1, then the integer gives the position of the
reference among the coefficients and the name gives the coefficient name. A non-named
numeric value of \code{ref} only works if the x-axis is also numeric (which can happen
in \code{iplot}).}

\item{ref.line}{Logical or numeric, default is "auto", whose behavior depends
on the situation. It is \code{TRUE} only if: i) interactions are plotted, ii) the
x values are numeric and iii) a reference is found. If \code{TRUE}, then a vertical
line is drawn at the level of the reference value. Otherwise, if numeric a vertical
line will be drawn at that specific value.}

\item{ref.line.par}{List. Parameters of the vertical line on the reference. The
default values are: \code{col = "black"} and \code{lty = 2}. You can add any graphical
parameter that will be passed to \code{\link[graphics:abline]{graphics::abline}}.
Eg: \code{ref.line.par = list(lty = 1, lwd = 3)}.}

\item{lab.cex}{The size of the labels of the coefficients. Default is missing.
It is automatically set by an internal algorithm which can go as low as \code{lab.min.cex}
(another argument).}

\item{lab.min.cex}{The minimum size of the coefficients labels, as set by the
internal algorithm. Default is 0.85.}

\item{lab.max.mar}{The maximum size the left margin can take when trying to fit
the coefficient labels into it (only when \code{horiz = TRUE}). This is used in the
internal algorithm fitting the coefficient labels. Default is \code{0.25}.}

\item{lab.fit}{The method to fit the coefficient labels into the plotting region
(only when \code{horiz = FALSE}). Can be \code{"auto"} (the default), \code{"simple"}, \code{"multi"}
or \code{"tilted"}. If \code{"simple"}, then the classic axis is drawn. If \code{"multi"}, then
the coefficient labels are fit horizontally across several lines, such that they
don't collide. If \code{"tilted"}, then the labels are tilted. If \code{"auto"}, an automatic
choice between the three is made.}

\item{xlim.add}{A numeric vector of length 1 or 2. It represents an extension
factor of xlim, in percentage. Eg: \code{xlim.add = c(0, 0.5)} extends \code{xlim} of 50\%
on the right. If of length 1, positive values represent the right, and negative
values the left (Eg: \code{xlim.add = -0.5} is equivalent to \code{xlim.add = c(0.5, 0)}).}

\item{ylim.add}{A numeric vector of length 1 or 2. It represents an extension
factor of ylim, in percentage. Eg: \code{ylim.add = c(0, 0.5)} extends \code{ylim} of 50\%
on the top. If of length 1, positive values represent the top, and negative values
the bottom (Eg: \code{ylim.add = -0.5} is equivalent to \code{ylim.add = c(0.5, 0)}).}

\item{only.params}{Logical, default is \code{FALSE}. If \code{TRUE} no graphic is displayed,
only the values of \code{x} and \code{y} used in the plot are returned.}

\item{sep}{The distance between two estimates -- only when argument \code{object}
is a list of estimation results.}

\item{as.multiple}{Logical: default is \code{FALSE}. Only when \code{object} is a single
estimation result: whether each coefficient should have a different color, line
type, etc. By default they all get the same style.}

\item{bg}{Background color for the plot. By default it is white.}

\item{group}{A list, default is missing. Each element of the list reports the
coefficients to be grouped while the name of the element is the group name. Each
element of the list can be either: i) a character vector of length 1, ii) of
length 2, or ii) a numeric vector. If equal to: i) then it is interpreted as
a pattern: all element fitting the regular expression will be grouped (note that
you can use the special character "^^" to clean the beginning of the names, see
example), if ii) it corresponds to the first and last elements to be grouped,
if iii) it corresponds to the coefficients numbers to be grouped. If equal to
a character vector, you can use a percentage to tell the algorithm to look at
the coefficients before aliasing (e.g. \code{"\%varname"}). Example of valid uses:
\verb{group=list(group_name=\\"pattern\\")}, \verb{group=list(group_name=c(\\"var_start\\", \\"var_end\\"))},
\verb{group=list(group_name=1:2))}. See details.}

\item{group.par}{A list of parameters controlling the display of the group. The
parameters controlling the line are: \code{lwd}, \code{tcl} (length of the tick), \code{line.adj}
(adjustment of the position, default is 0), \code{tick} (whether to add the ticks),
\code{lwd.ticks}, \code{col.ticks}. Then the parameters controlling the text: \code{text.adj}
(adjustment of the position, default is 0), \code{text.cex}, \code{text.font}, \code{text.col}.}

\item{main}{The title of the plot. Default is \code{"Effect on __depvar__"}. You can
use the special variable \verb{__depvar__} to set the title (useful when you set the
plot default with \code{\link{setFixest_coefplot}}).}

\item{value.lab}{The label to appear on the side of the coefficient values. If
\code{horiz = FALSE}, the label appears in the y-axis. If \code{horiz = TRUE}, then it
appears on the x-axis. The default is equal to \code{"Estimate and __ci__ Conf. Int."},
with \verb{__ci__} a special variable giving the value of the confidence interval.}

\item{ylab}{The label of the y-axis, default is \code{NULL}. Note that if
\code{horiz = FALSE}, it overrides the value of the argument \code{value.lab}.}

\item{xlab}{The label of the x-axis, default is \code{NULL}. Note that if \code{horiz =  TRUE}, it overrides the value of the argument \code{value.lab}.}

\item{sub}{A subtitle, default is \code{NULL}.}

\item{i.select}{Integer scalar, default is 1. In \code{iplot}, used to select which
variable created with \code{i()} to select. Only used when there are several variables
created with \code{\link{i}}. This is an index, just try increasing numbers to hopefully
obtain what you want. Note that it works much better when the variables are "pure"
\code{i()} and not interacted with other variables. For example: \code{i(species, x1)}
is good while \code{i(species):x1} isn't. The latter will also work but the index
may feel weird in case there are many \code{i()} variables.}

\item{do_iplot}{Logical, default is \code{FALSE}. For internal use only.
If \code{TRUE}, then \code{iplot} is run instead of \code{coefplot}.}
}
\description{
This function plots the results of estimations (coefficients and confidence intervals).
The function \code{iplot} restricts the output to variables created with \code{\link{i}}, either
interactions with factors or raw factors.
}
\section{Functions}{
\itemize{
\item \code{iplot()}: Plots the coefficients generated with i()

}}
\section{Setting custom default values}{

The function \code{coefplot} dispose of many arguments to parametrize the plots. Most
of these arguments can be set once an for all using the function \code{\link{setFixest_coefplot}}.
See Example 3 below for a demonstration.
}

\section{iplot}{


The function \code{iplot} restricts \code{coefplot} to interactions or factors created
with the function \code{\link{i}}. Only \emph{one} of the i-variables will be plotted at a time.
If you have several i-variables, you can navigate through them with the \code{i.select} argument.

The argument \code{i.select} is an index that will go through all the i-variables.
It will work well if the variables are pure, meaning not interacted with other
variables. If the i-variables are interacted, the index may have an odd behavior
but will (in most cases) work all the same, just try some numbers up until you
(hopefully) obtain the graph you want.

Note, importantly, that interactions of two factor variables are (in general)
disregarded since they would require a 3-D plot to be properly represented.
}

\section{Mathematical expressions}{


You can add \code{\link[grDevices:plotmath]{plotmath}} mathematical expressions in the arguments
\code{main}, \code{sub}, \code{xlab}, or \code{ylab}. To do so, start the character string with an ampersand.
For example main = "&lambda^2".
}

\section{Arguments keep, drop and order}{

The arguments \code{keep}, \code{drop} and \code{order} use regular expressions. If you are not aware
of regular expressions, I urge you to learn it, since it is an extremely powerful way
to manipulate character strings (and it exists across most programming languages).

For example drop = "Wind" would drop any variable whose name contains "Wind". Note that
variables such as "Temp:Wind" or "StrongWind" do contain "Wind", so would be dropped.
To drop only the variable named "Wind", you need to use
\code{drop = "^Wind$"} (with "^" meaning beginning, resp. "$" meaning end,
of the string => this is the language of regular expressions).

Although you can combine several regular expressions in a single character
string using pipes, \code{drop} also accepts a vector of regular expressions.

You can use the special character "!" (exclamation mark) to reverse the effect
of the regular expression (this feature is specific to this function).
For example \code{drop = "!Wind"} would drop any variable that does not contain "Wind".

By default, the regular expressions are checked against the variables after
they have been renamed with the dictionary (argument \code{dict}).
You can use the \verb{*_raw} versions of drop/keep/order to apply the regular
expressions on the original variable names.
Note that alternatively you can use the special character "\%" (percentage) at the
beginning of drop/keep/order's regular expressions to refer to the original variable name.
For example, you have a variable named \code{"Month6"},
and use a dictionary \code{dict = c(Month6="June")}.
Thus the variable will be displayed as \code{"June"}.
If you want to delete that variable, you can use either \code{drop="June"}, \code{drop_raw="Month6"},
or \code{drop="\%Month6"}.

The argument \code{order} takes in a vector of regular expressions, the order will follow the
elements of this vector. The vector gives a list of priorities,
on the left the elements with highest priority.
For example, order = c("Wind", "!Inter", "!Temp") would give highest priorities to
the variables containing "Wind" (which would then appear first),
second highest priority is the variables not containing "Inter", last,
with lowest priority, the variables not containing "Temp".
If you had the following variables: (Intercept), Temp:Wind, Wind, Temp you
would end up with the following order: Wind, Temp:Wind, Temp, (Intercept).
}

\examples{

#
# Example 1: Stacking two sets of results on the same graph
#

# Estimation on Iris data with one fixed-effect (Species)
# + we cluster the standard-errors
est = feols(Petal.Length ~ Petal.Width + Sepal.Width | Species, 
            iris, vcov = "cluster")

# Now with "regular" standard-errors
est_std = summary(est, vcov = "iid")

# You can plot the two results at once
coefplot(est, est_std)

# You could also use the argument vcov
coefplot(est, vcov = list("cluster", "iid"))

# Alternatively, you can use the argument x.shift
# to do it sequentially:

# First graph with clustered standard-errors
coefplot(est, x.shift = -.2)

# 'x.shift' was used to shift the coefficients to the left.

# Second set of results: this time with
#  standard-errors that are not clustered.
coefplot(est, vcov = "iid", x.shift = .2,
         add = TRUE, col = 2, ci.lty = 2, pch = 15)

legend("topright", col = 1:2, pch = 20, lwd = 1, lty = 1:2,
       legend = c("Clustered", "IID"), title = "Standard-Errors")


#
# Example 2: Interactions
#


# Now we estimate and plot the "yearly" treatment effects

data(base_did)
base_inter = base_did

# We interact the variable 'period' with the variable 'treat'
est_did = feols(y ~ x1 + i(period, treat, 5) | id + period, base_inter)

# In the estimation, the variable treat is interacted
#  with each value of period but 5, set as a reference

# coefplot will show all the coefficients:
coefplot(est_did)

# Note that the grouping of the coefficients is due to 'group = "auto"'

# If you want to keep only the coefficients
# created with i() (ie the interactions), use iplot
iplot(est_did)

# We can see that the graph is different from before:
#  - only interactions are shown,
#  - the reference is present,
# => this is fully flexible

iplot(est_did, ref.line = FALSE, pt.join = TRUE)


#
# What if the interacted variable is not numeric?

# Let's create a "month" variable
all_months = c("aug", "sept", "oct", "nov", "dec", "jan",
               "feb", "mar", "apr", "may", "jun", "jul")
base_inter$period_month = all_months[base_inter$period]

# The new estimation
est = feols(y ~ x1 + i(period_month, treat, "oct") | id+period, base_inter)
# Since 'period_month' of type character, coefplot sorts it
iplot(est)

# To respect a plotting order, use a factor
base_inter$month_factor = factor(base_inter$period_month, levels = all_months)
est = feols(y ~ x1 + i(month_factor, treat, "oct") | id + period, base_inter)
iplot(est)


#
# Example 3: Setting defaults
#

# coefplot has many arguments, which makes it highly flexible.
# If you don't like the default style of coefplot. No worries,
# you can set *your* default by using the function
# setFixest_coefplot()

dict = c("Petal.Length"="Length (Petal)", "Petal.Width"="Width (Petal)",
         "Sepal.Length"="Length (Sepal)", "Sepal.Width"="Width (Sepal)")

setFixest_coefplot(ci.col = 2, pt.col = "darkblue", ci.lwd = 3,
                   pt.cex = 2, pt.pch = 15, ci.width = 0, dict = dict)

est = feols(Petal.Length ~ Petal.Width + Sepal.Length +
                Sepal.Width + i(Species), iris)

# And that's it
coefplot(est)

# You can set separate default values for iplot
setFixest_coefplot("iplot", pt.join = TRUE, pt.join.par = list(lwd = 2, lty = 2))
iplot(est)

# To reset to the default settings:
setFixest_coefplot("all", reset = TRUE)
coefplot(est)

#
# Example 4: group + cleaning
#

# You can use the argument group to group variables
# You can further use the special character "^^" to clean
#  the beginning of the coef. name: particularly useful for factors

est = feols(Petal.Length ~ Petal.Width + Sepal.Length +
                Sepal.Width + Species, iris)

# No grouping:
coefplot(est)

# now we group by Sepal and Species
coefplot(est, group = list(Sepal = "Sepal", Species = "Species"))

# now we group + clean the beginning of the names using the special character ^^
coefplot(est, group = list(Sepal = "^^Sepal.", Species = "^^Species"))


}
\seealso{
See \code{\link{setFixest_coefplot}} to set the default values of \code{coefplot}, and the estimation
functions: e.g. \code{\link{feols}}, \code{\link[=feglm]{fepois}}, \code{\link{feglm}}, \code{\link[=femlm]{fenegbin}}.
}
\author{
Laurent Berge
}
