% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\name{degrees_freedom}
\alias{degrees_freedom}
\title{Gets the degrees of freedom of a \code{fixest} estimation}
\usage{
degrees_freedom(
  x,
  type,
  vars = NULL,
  se = NULL,
  cluster = NULL,
  dof = NULL,
  stage = 2
)
}
\arguments{
\item{x}{A \code{fixest} estimation.}

\item{type}{Integer, equal to 1 or 2 (default). If 1, then the number of regressors is returned. If 2, then it is the "residuals degree of freedom", i.e. the number of observations minus the number of regressors. Note that the latter value is affected by how the VCOV of \code{x} is computed, in particular when the VCOV is clustered (see help of \code{\link[fixest]{dof}}, in particular the \code{t.df} argument).}

\item{vars}{A vector of variable names, reflecting regressors. This is optional. If provided, then \code{type} is set to 1 by default and the number of regressoors contained in \code{vars} is returned. This is only useful in the presence of collinearity and we want a subset of the regressors only. (Mostly for internal use.)}

\item{se}{Character scalar. Which kind of standard error should be computed: \dQuote{standard}, \dQuote{hetero}, \dQuote{cluster}, \dQuote{twoway}, \dQuote{threeway} or \dQuote{fourway}? By default if there are clusters in the estimation: \code{se = "cluster"}, otherwise \code{se = "standard"}. Note that this argument can be implicitly deduced from the argument \code{cluster}.}

\item{cluster}{Tells how to cluster the standard-errors (if clustering is requested). Can be either a list of vectors, a character vector of variable names, a formula or an integer vector. Assume we want to perform 2-way clustering over \code{var1} and \code{var2} contained in the data.frame \code{base} used for the estimation. All the following \code{cluster} arguments are valid and do the same thing: \code{cluster = base[, c("var1", "var2")]}, \code{cluster = c("var1", "var2")}, \code{cluster = ~var1+var2}. If the two variables were used as clusters in the estimation, you could further use \code{cluster = 1:2} or leave it blank with \code{se = "twoway"} (assuming \code{var1} [resp. \code{var2}] was the 1st [res. 2nd] cluster). You can interact two variables using \code{^} with the following syntax: \code{cluster = ~var1^var2} or \code{cluster = "var1^var2"}.}

\item{dof}{An object of class \code{dof.type} obtained with the function \code{\link[fixest]{dof}}. Represents how the degree of freedom correction should be done.You must use the function \code{\link[fixest]{dof}} for this argument. The arguments and defaults of the function \code{\link[fixest]{dof}} are: \code{adj = TRUE}, \code{fixef.K="nested"}, \code{cluster.adj = TRUE}, \code{cluster.df = "conventional"}, \code{t.df = "conventional"}, \code{fixef.force_exact=FALSE)}. See the help of the function \code{\link[fixest]{dof}} for details.}

\item{stage}{Either 1 or 2. Only concerns IV regressions, which stage to look at.}
}
\description{
Simple utility to extract the degrees of freedom from a \code{fixest} estimation.
}
\examples{

# First: an estimation

base = iris
names(base) = c("y", "x1", "x2", "x3", "species")
est = feols(y ~ x1 + x2 | species, base)

# "Normal" standard-errors (SE)
est_standard = summary(est, se = "st")

# Clustered SEs
est_clustered = summary(est, se = "clu")

# The different degrees of freedom

# => different type 1 DoF (because of the clustering)
degrees_freedom(est_standard, type = 1)
degrees_freedom(est_clustered, type = 1) # fixed-effects are excluded

# => different type 2 DoF (because of the clustering)
degrees_freedom(est_standard, type = 2) # => equivalent to the df.residual from lm
degrees_freedom(est_clustered, type = 2)



}
