% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{deviance.fixest}
\alias{deviance.fixest}
\title{Extracts the deviance of a fixest estimation}
\usage{
\method{deviance}{fixest}(object, ...)
}
\arguments{
\item{object}{A `fixest` object.}

\item{...}{Not currently used.}
}
\value{
Returns a numeric scalar equal to the deviance.
}
\description{
Returns the deviance from a `fixest` estimation.
}
\examples{

est <- feols(Petal.Length ~ Petal.Width, iris)
deviance(est)

est_pois <- fepois(Petal.Length ~ Petal.Width, iris)
deviance(est_pois)

}
\seealso{
[`feols`], [`fepois`][fixest2::feglm], [`feglm`], [`fenegbin`][fixest2::femlm], [`feNmlm`].
}
