% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{deviance.fixest2}
\alias{deviance.fixest2}
\title{Extracts the deviance of a fixest2 estimation}
\usage{
\method{deviance}{fixest2}(object, ...)
}
\arguments{
\item{object}{A \code{fixest2} object.}

\item{...}{Not currently used.}
}
\value{
Returns a numeric scalar equal to the deviance.
}
\description{
Returns the deviance from a \code{fixest2} estimation.
}
\examples{

est <- feols(Petal.Length ~ Petal.Width, iris)
deviance(est)

est_pois <- fepois(Petal.Length ~ Petal.Width, iris)
deviance(est_pois)

}
\seealso{
\code{\link{feols}}, \code{\link[=feglm]{fepois}}, \code{\link{feglm}}, \code{\link[=femlm]{fenegbin}}, \code{\link{feNmlm}}.
}
