% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_test.R
\name{.ar_exact_extra_args}
\alias{.ar_exact_extra_args}
\title{Identify additional arguments beyond those supported by the exact AR CI}
\usage{
.ar_exact_extra_args(object)
}
\description{
Identify additional arguments beyond those supported by the exact AR CI
}
\keyword{internal}
