% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_test.R
\name{.ar_test_core}
\alias{.ar_test_core}
\title{Internal core function for Anderson-Rubin test}
\usage{
.ar_test_core(object, beta0, vcov = NULL, orig_call = NULL, ...)
}
\arguments{
\item{object}{A \code{fixest} object obtained from \code{\link{feols}} with an IV specification.}

\item{beta0}{Numeric vector. The hypothesized values for the endogenous variable
coefficients under the null hypothesis.}

\item{vcov}{Versatile argument to specify the VCOV.}

\item{...}{Additional arguments passed to \code{\link{feols}}.}
}
\value{
A list containing:
\item{stat}{The test statistic (F or Chi-squared depending on vcov).}
\item{df1}{Numerator degrees of freedom.}
\item{df2}{Denominator degrees of freedom (NA for Chi-squared).}
\item{pvalue}{The p-value.}
\item{dist}{Distribution used: "F" or "Chi-squared".}
\item{vcov_type}{The VCOV type used.}
\item{endo_names}{Names of endogenous variables.}
\item{inst_names}{Names of instruments.}
\item{inst_tested}{Instrument coefficients actually tested.}
}
\description{
This is an internal function that computes the AR test statistic and p-value.
It is called by both ar_test() and ar_confint().
}
\keyword{internal}
