% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_test.R
\name{.print_ar_ci}
\alias{.print_ar_ci}
\title{Internal function to print AR confidence interval}
\usage{
.print_ar_ci(ci, endo_name)
}
\description{
Internal function to print AR confidence interval
}
\keyword{internal}
