% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{estfun_reexported}
\alias{estfun_reexported}
\alias{estfun}
\title{Functions exported from \pkg{sandwich} to implement \pkg{fixest} methods}
\description{
The package \pkg{fixest} does not use \code{estfun} from \pkg{sandwich}, but this method has been implemented to allow users to leverage the variances from \pkg{sandwich}.
}
\details{
\itemize{
\item Here is the help from package \pkg{sandwich}: \code{\link[sandwich:estfun]{estfun}}. The help from package \pkg{fixest} is here: \code{\link[fixest]{estfun.fixest}}.
}
}
\keyword{internal}
