% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{fdim}
\alias{fdim}
\title{Formatted dimension}
\usage{
fdim(x)
}
\arguments{
\item{x}{An R object, usually a data.frame (but can also be a vector).}
}
\value{
It does not return anything, the output is directly printed on the console.
}
\description{
Prints the dimension of a data set, in an user-readable way
}
\examples{

fdim(iris)

fdim(iris$Species)



}
\author{
Laurent Berge
}
