% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated_funs.R
\name{fixest-deprecated}
\alias{fixest-deprecated}
\alias{did_estimate_yearly_effects-deprecated}
\alias{did_plot_yearly_effects-deprecated}
\alias{errbar-deprecated}
\title{Deprecated functions in package \pkg{fixest}.}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\section{\code{did_estimate_yearly_effects}}{

From \code{fixest} version 0.3.0 onwards, estimating yearly effects with a simple interaction of the type \code{var::year(ref)} in the formula is preferred (see details in \code{\link[fixest]{feols}}). You can then plot the yearly treatment with the function \code{\link[fixest]{coefplot}}. You have examples detailed in \code{\link[fixest]{coefplot}}.
}

\section{\code{did_plot_yearly_effects}}{

From \code{fixest} version 0.3.0 onwards, estimating yearly effects with a simple interaction of the type \code{var::year(ref)} in the formula is preferred (see details in \code{\link[fixest]{feols}}). You can then plot the yearly treatment with the function \code{\link[fixest]{coefplot}}. You have examples detailed in \code{\link[fixest]{coefplot}}.
}

\section{\code{errbar}}{

The function \code{errbar} is now deprecated and will be removed in the near future. Use \code{\link[fixest]{coefplot}} instead, which dispose of much more possibilities..
}

\keyword{internal}
