% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfuns.R
\docType{data}
\name{fulton}
\alias{fulton}
\title{Fulton Fish Market data}
\format{
An object of class \code{data.frame} with 97 rows and 12 columns.
}
\source{
https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041
}
\usage{
data(fulton)

`fulton` is a data frame with 97 observations and 12 variables named `t`, `day`, `price`, `qty`, `speed2`, `wave2`, `speed3`, `wave3`, `price_asian`, `price_white`, `qty_asian`, `qty_white`. 
Each row is a recording of the Fulton fish market sales on a given day.

* t: Time-trend
* day: Day of the week 
* price: Average price of fish (calculated as `(qty_white * price_white + qty_asian * price_asian) / (qty_white + qty_asian)`)
* qty: Quantity of fish sold (calculated as `qty_white + qty_asian`)
* speed2: Wind speeds (minimum of past 2 days)
* wave2:  Maximum wave height (average of last 2 days)
* speed3: Wind speed (3 day lag)
* wave3:  Maximum wave height (average of last 3 and 4 day lag)
* price_asian: Average price of fish sold to Asian customers
* price_white: Average price of fish sold to White customers
* qty_asian: Quantity of fish sold to Asian customers
* qty_white: Quantity of fish sold to White customers
}
\description{
This dataset has been taken from Jeff Wooldridge's textbook.
A modified version that appears in the \code{wooldridge} package.
}
\details{
Source: K Graddy (1995), “Testing for Imperfect Competition at the Fulton Fish Market,” RAND Journal of Economics 26, 75-92.
}
\keyword{datasets}
