% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCOV.R
\name{getFixest_ssc}
\alias{getFixest_ssc}
\title{Gets the default values for the small sample correction}
\usage{
getFixest_ssc(vcov_name = NULL)
}
\arguments{
\item{vcov_name}{A character scalar naming the type of VCOV, or \code{NULL}. It should give
the name of a valid \code{fixest} VCOV (see \code{\link{vcov.fixest}}). If so, an object of class
\code{ssc_type} is returned.
If \code{NULL}, the list of all small sample corrections (of class \code{ssc_type})
which have been set by the user is returned.}
}
\value{
If \code{vcov_name} is a valid \code{fixest} VCOV name, an object of class \code{ssc_type} is returned,
it corresponds to the default small sample correction (SSC) for this VCOV.

If \code{NULL}, all the SSCs that have been set by the user with \code{\link{setFixest_ssc}} are returned.
}
\description{
Gets the default small sample correction, of class \code{ssc_type}, for specific \code{fixest} VCOVs
}
\examples{

# default SSC for iid
getFixest_ssc("iid")

# in a fresh session, the default for Newey West VCOVs is different:
getFixest_ssc("NW")

}
\seealso{
\code{\link{vcov.fixest}}, \code{\link{ssc}}, \code{\link{setFixest_ssc}}
}
