% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\name{osize}
\alias{osize}
\alias{print.osize}
\title{Formatted object size}
\usage{
osize(x)

\method{print}{osize}(x, ...)
}
\arguments{
\item{x}{Any R object.}

\item{...}{Not currently used.}
}
\value{
Returns a character scalar.
}
\description{
Tools that returns a formatted object size, where the appropriate unit is automatically chosen.
}
\examples{

osize(iris)

data(trade)
osize(trade)


}
