% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{plot.fixest}
\alias{plot.fixest}
\title{plot methods for \code{fixest} and \code{fixest_multi} objects}
\usage{
\method{plot}{fixest}(
  x,
  vcov = NULL,
  add = FALSE,
  horiz = FALSE,
  do_iplot = NULL,
  zero = TRUE,
  zero.par = TRUE,
  dict = NULL,
  keep = NULL,
  drop = NULL,
  order = NULL,
  ci.width = "1\%",
  ci_level = 0.95,
  plot_prms = list(),
  ylim = NULL,
  xlim = NULL,
  pch = c(20, 17, 15, 21, 24, 22),
  col = 1:8,
  cex = 1,
  lty = 1,
  lwd = 1,
  pt.pch = pch,
  pt.bg = NULL,
  pt.cex = cex,
  pt.col = col,
  ci.col = col,
  pt.lwd = lwd,
  ci.lwd = lwd,
  ci.lty = lty,
  main = "Effect on __depvar__",
  value.lab = "Estimate and __ci__ Conf. Int.",
  ylab = NULL,
  xlab = NULL,
  sub = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{fixest} estimation, for example from \code{\link{feols}}.}

\item{vcov}{Versatile argument to specify the VCOV.
In general, it is either a character scalar equal to a VCOV type, either a formula of the form:
vcov_type ~ variables. The VCOV types implemented are: "iid", "hetero" (or "HC1"),
"cluster", "twoway", "NW" (or "newey_west"), "DK" (or "driscoll_kraay"), and "conley".
It also accepts object from vcov_cluster, vcov_NW, NW, vcov_DK, DK, vcov_conley and conley.
It also accepts covariance matrices computed externally.
Finally it accepts functions to compute the covariances.
See the vcov documentation in the vignette.

You can pass several VCOVs (as above) if you nest them into a list.
If the number of VCOVs equals the number of models, each VCOV is mapped to the appropriate model.
If there is one model and several VCOVs, or if the first element of the list is equal to
\code{"each"} or \code{"times"}, then the estimations will be replicated and the results
for each estimation and each VCOV will be reported.}

\item{add}{Default is \code{FALSE}, if the intervals are to be added to an existing
graph. Note that if it is the case, then the argument \code{x} MUST be numeric.}

\item{horiz}{A logical scalar, default is \code{FALSE}. Whether to display the confidence
intervals horizontally instead of vertically.}

\item{do_iplot}{Logical, default is \code{FALSE}. For internal use only.
If \code{TRUE}, then \code{iplot} is run instead of \code{coefplot}.}

\item{zero}{Logical scalar, default is \code{TRUE}. Whether the 0 should be displayed
in the limits of the y-axis.
Note that you can set how this zero line looks like with the argument \code{zero.par}.}

\item{zero.par}{A named list of graphical parameters or a logical scalar.
This argument is a list containing the graphical parameters used to draw the zero-line.
The default value is \code{list(col = "black", lwd = 1)} (it's the same if \code{TRUE}).
Set it to \code{FALSE} to turn off the special emphasis of the zero line.
You can add any graphical parameter that will be passed
to \code{\link[graphics:abline]{graphics::abline}}. Example: \code{zero.par = list(col = "darkblue", lwd = 3)}.}

\item{dict}{A named character vector or a logical scalar. It changes the original variable names
to the ones contained in the \code{dict}ionary. E.g. to change the variables named \code{a} and \code{b3} to
(resp.) \dQuote{$log(a)$} and to \dQuote{$bonus^3$}, use \code{dict=c(a="$log(a)$",b3="$bonus^3$")}.
By default, it is equal to \code{getFixest_dict()}, a default dictionary which can be set with
\code{\link{setFixest_dict}}. You can use \code{dict = FALSE} to disable it. By default \code{dict} modifies the
entries in the global dictionary, to disable this behavior, use "reset" as the first element
(ex: \code{dict=c("reset", mpg="Miles per gallon")}).}

\item{keep}{Character vector. This element is used to display only a subset of variables. This
should be a vector of regular expressions (see \code{\link[base:regex]{base::regex}} help for more info). Each
variable satisfying any of the regular expressions will be kept. This argument is applied post
aliasing (see argument \code{dict}).
Use the argument \code{keep_raw} for the same effect before aliasing.

Example: you have the variable \code{x1} to \code{x55} and want to display
only \code{x1} to \code{x9}, then you could use \code{keep = "x[[:digit:]]$"}. If the first character is an
exclamation mark, the effect is reversed (e.g. keep = "!Constant" means: every variable that
does not contain \dQuote{Constant} is kept). See details.}

\item{drop}{Character vector. This element is used if some variables are not to be displayed.
This should be a vector of regular expressions (see \code{\link[base:regex]{base::regex}} help for more info). Each
variable satisfying any of the regular expressions will be discarded. This argument is applied
post aliasing (see argument \code{dict}).
Use the argument \code{drop_raw} for the same effect before aliasing.

Example: you have the variable \code{x1} to \code{x55} and want to
display only \code{x1} to \code{x9}, then you could use \verb{drop = "x[[:digit:]]\{2\}}". If the first character
is an exclamation mark, the effect is reversed (e.g. drop = "!Constant" means: every variable
that does not contain \dQuote{Constant} is dropped). See details.}

\item{order}{Character vector. This element is used if the user wants the variables to be
ordered in a certain way. This should be a vector of regular expressions (see \code{\link[base:regex]{base::regex}}
help for more info). The variables satisfying the first regular expression will be placed first,
then the order follows the sequence of regular expressions. This argument is applied post
aliasing (see argument \code{dict}). Use the argument \code{order_raw} for the same effect before aliasing.

Example: you have the following variables: \code{month1} to \code{month6},
then \code{x1} to \code{x5}, then \code{year1} to \code{year6}. If you want to display first the x's, then the
years, then the months you could use: \code{order = c("x", "year")}. If the first character is an
exclamation mark, the effect is reversed (e.g. order = "!Constant" means: every variable that
does not contain \dQuote{Constant} goes first).  See details.}

\item{ci.width}{The width of the extremities of the confidence intervals. Default is \code{0.1}.}

\item{ci_level}{Scalar between 0 and 1: the level of the CI. By default it is equal to 0.95.}

\item{plot_prms}{A named list. It may contain additionnal parameters to be passed
to the plot.}

\item{ylim}{Numeric vector of length 2 which gives the limits of the plotting region for
the y-axis. The default is \code{NULL}, which means that it is automatically defined.
Use the argument \code{ylim.add} to simply increase or decrease the default limits.}

\item{xlim}{Numeric vector of length 2 which gives the limits of the plotting region for
the x-axis. The default is \code{NULL}, which means that it is automatically defined.
Use the argument \code{xlim.add} to simply increase or decrease the default limits.}

\item{pch}{The pch of the coefficient estimates. Default is 1 (circle).
This is an alias to the argument \code{pt.pch}.}

\item{col}{The color of the points and the confidence intervals. Default is 1
("black"). Note that you can set the colors separately for each of them
with \code{pt.col} and \code{ci.col}.}

\item{cex}{Numeric, default is 1. Expansion factor for the points}

\item{lty}{The line type of the confidence intervals. Default is 1.
This is an alias to the argument \code{ci.lty}.}

\item{lwd}{General line width. Default is 1.}

\item{pt.pch}{The pch of the coefficient estimates. Default is 1 (circle).}

\item{pt.bg}{The background color of the point estimate (when the \code{pt.pch} is
in 21 to 25). Defaults to NULL.}

\item{pt.cex}{The size of the coefficient estimates. Default is the other argument \code{cex}.}

\item{pt.col}{The color of the coefficient estimates. Default is equal to the argument \code{col}.}

\item{ci.col}{The color of the confidence intervals. Default is equal to the argument \code{col}.}

\item{pt.lwd}{The line width of the coefficient estimates. Default is equal to
the other argument \code{lwd}.}

\item{ci.lwd}{The line width of the confidence intervals. Default is equal to
the other argument \code{lwd}.}

\item{ci.lty}{The line type of the confidence intervals. Default is 1.}

\item{main}{The title of the plot. Default is \code{"Effect on __depvar__"}. You can
use the special variable \verb{__depvar__} to set the title (useful when you set the
plot default with \code{\link{setFixest_coefplot}}).}

\item{value.lab}{The label to appear on the side of the coefficient values. If
\code{horiz = FALSE}, the label appears in the y-axis. If \code{horiz = TRUE}, then it
appears on the x-axis. The default is equal to \code{"Estimate and __ci__ Conf. Int."},
with \verb{__ci__} a special variable giving the value of the confidence interval.}

\item{ylab}{The label of the y-axis, default is \code{NULL}. Note that if
\code{horiz = FALSE}, it overrides the value of the argument \code{value.lab}.}

\item{xlab}{The label of the x-axis, default is \code{NULL}. Note that if \code{horiz =  TRUE}, it overrides the value of the argument \code{value.lab}.}

\item{sub}{A subtitle, default is \code{NULL}.}

\item{...}{Other arguments to be passed to \code{\link{coefplot}}.}
}
\value{
It returns invisibly the data used to create the graph.
}
\description{
Plot method reporting the coefficient estimates and their confidence intervals.
This is a wrapper to the more complete functions \code{\link{coefplot}} and \code{\link{iplot}}.
}
\details{
By default \code{plot.fixest} runs \code{\link{coefplot}} unless the estimation includes
the function \code{\link{sunab}}, in which case it uses \code{\link{iplot}}.

The switch to \code{iplot} can be made with the argument \code{do_iplot = TRUE}.
}
\section{Arguments keep, drop and order}{

The arguments \code{keep}, \code{drop} and \code{order} use regular expressions. If you are not aware
of regular expressions, I urge you to learn it, since it is an extremely powerful way
to manipulate character strings (and it exists across most programming languages).

For example drop = "Wind" would drop any variable whose name contains "Wind". Note that
variables such as "Temp:Wind" or "StrongWind" do contain "Wind", so would be dropped.
To drop only the variable named "Wind", you need to use
\code{drop = "^Wind$"} (with "^" meaning beginning, resp. "$" meaning end,
of the string => this is the language of regular expressions).

Although you can combine several regular expressions in a single character
string using pipes, \code{drop} also accepts a vector of regular expressions.

You can use the special character "!" (exclamation mark) to reverse the effect
of the regular expression (this feature is specific to this function).
For example \code{drop = "!Wind"} would drop any variable that does not contain "Wind".

By default, the regular expressions are checked against the variables after
they have been renamed with the dictionary (argument \code{dict}).
You can use the \verb{*_raw} versions of drop/keep/order to apply the regular
expressions on the original variable names.
Note that alternatively you can use the special character "\%" (percentage) at the
beginning of drop/keep/order's regular expressions to refer to the original variable name.
For example, you have a variable named \code{"Month6"},
and use a dictionary \code{dict = c(Month6="June")}.
Thus the variable will be displayed as \code{"June"}.
If you want to delete that variable, you can use either \code{drop="June"}, \code{drop_raw="Month6"},
or \code{drop="\%Month6"}.

The argument \code{order} takes in a vector of regular expressions, the order will follow the
elements of this vector. The vector gives a list of priorities,
on the left the elements with highest priority.
For example, order = c("Wind", "!Inter", "!Temp") would give highest priorities to
the variables containing "Wind" (which would then appear first),
second highest priority is the variables not containing "Inter", last,
with lowest priority, the variables not containing "Temp".
If you had the following variables: (Intercept), Temp:Wind, Wind, Temp you
would end up with the following order: Wind, Temp:Wind, Temp, (Intercept).
}

\examples{

#
# Single estimation
#

est = feols(Ozone ~ Temp + Solar.R, airquality)
plot(est)

# focus only on the variables
plot(est, drop = "Cons")

#
# Multiple estimations
#

est_mult = feols(Ozone ~ csw(Temp, Solar.R, Wind), airquality)
plot(est_mult, drop = "Const")

#
# DiD estimation: Sun & Abraham
#

data(base_stagg)
# The DiD estimation
res_sunab = feols(y ~ x1 + sunab(year_treated, year) | id + year, base_stagg)
plot(res_sunab)




}
