% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fixest_vcov}
\alias{print.fixest_vcov}
\title{A print facility for the VCOVs from \code{fixest} objects.}
\usage{
\method{print}{fixest_vcov}(x, ...)
}
\arguments{
\item{x}{A \code{fixest_vcov} object, obtained from \code{\link{vcov.fixest}}.}

\item{...}{Not used.}
}
\value{
This function does not return anything.
}
\description{
Prints a VCOV obtained from \code{fixest}, on top of a regular matrix display, its main use is to:
i) report how the VCOV was computed, and ii) hide the information on attributes
}
\examples{

# 1) fixest estimation
est = feols(Petal.Length ~ Sepal.Width, iris)

# 2) print the VCOV; this method hides the attributes
vcov(est)

# 3) showing the attributes if needed
attributes(vcov(est))


}
\author{
Laurent Berge
}
