% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCOV.R
\name{print.ssc_type}
\alias{print.ssc_type}
\title{Method to print the type of small sample correction}
\usage{
\method{print}{ssc_type}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ssc_type}, obtained from the function \code{\link{ssc}}.}

\item{...}{Not currently used.}
}
\value{
This function does not return anything.
}
\description{
Pretty print for the small sample correction obtained from \code{\link{ssc}}
}
\examples{

# ssc() is just a list
unclass(ssc())

# print of ssc, much more compactly
ssc()

}
\seealso{
\code{\link{vcov.fixest}}, \code{\link{ssc}}, \code{\link{setFixest_ssc}}, \code{\link{getFixest_ssc}}
}
