% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_env.R
\name{reshape_env}
\alias{reshape_env}
\title{\itemize{
\item rm_0s: whether to remove the variables that are only equal to 0
}}
\usage{
reshape_env(
  env,
  obs2keep = NULL,
  lhs = NULL,
  rhs = NULL,
  assign_lhs = TRUE,
  assign_rhs = TRUE,
  fml_linear = NULL,
  fml_fixef = NULL,
  fml_iv_endo = NULL,
  check_lhs = FALSE,
  assign_fixef = FALSE,
  rm_0s = FALSE
)
}
\description{
\itemize{
\item rm_0s: whether to remove the variables that are only equal to 0
}
}
