% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\docType{data}
\name{setFixest_dof}
\alias{setFixest_dof}
\alias{getFixest_dof}
\title{Sets the default type of DoF correction in summary/vcov.fixest}
\format{
An object of class \code{function} of length 1.
}
\usage{
setFixest_dof(dof.type = dof())

getFixest_dof
}
\arguments{
\item{dof.type}{An object of class \code{dof.type} obtained with the function \code{\link[fixest]{dof}}.}
}
\value{
The function \code{getFixest_dof} returns a \code{dof.type} object.
}
\description{
Sets or gets the  default type of DOF correction in \code{\link[fixest]{summary.fixest}} and \code{\link[fixest]{vcov.fixest}}
}
\examples{

\dontrun{

# If you never want DoF correction when computing the vcov
# of fixest object:

setFixest_dof(dof(adj = 0, cluster.adj = FALSE))

}


}
\seealso{
\code{\link[fixest]{dof}}
}
\author{
Laurent Berge
}
\keyword{datasets}
