% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfuns.R
\name{setFixest_estimation}
\alias{setFixest_estimation}
\alias{getFixest_estimation}
\title{Default arguments for fixest estimations}
\usage{
setFixest_estimation(
  data = NULL,
  panel.id = NULL,
  fixef.rm = "perfect_fit",
  fixef.tol = 1e-06,
  fixef.iter = 10000,
  collin.tol = 1e-10,
  lean = FALSE,
  verbose = 0,
  warn = TRUE,
  fixef.keep_names = NULL,
  demeaned = FALSE,
  mem.clean = FALSE,
  glm.iter = 25,
  glm.tol = 1e-08,
  demeaner = NULL,
  data.save = FALSE,
  reset = FALSE
)

getFixest_estimation()
}
\arguments{
\item{data}{A data.frame containing the necessary variables to run the model.
The variables of the non-linear right hand side of the formula are identified
with this \code{data.frame} names. Can also be a matrix.}

\item{panel.id}{The panel identifiers. Can either be: i) a one sided formula
(e.g. \code{panel.id = ~id+time}), ii) a character vector of length 2
(e.g. \code{panel.id=c('id', 'time')}, or iii) a character scalar of two variables
separated by a comma (e.g. \code{panel.id='id,time'}). Note that you can combine variables
with \code{^} only inside formulas (see the dedicated section in \code{\link{feols}}).}

\item{fixef.rm}{Can be equal to "perfect_fit" (default), "singletons", "infinite_coef"
or "none".

This option controls which observations should be removed prior to the estimation.
If "singletons", fixed-effects associated to a single observation are removed
(since they perfectly explain it).

The value "infinite_coef" only works with GLM families with limited left hand sides (LHS)
and exponential link.
For instance the Poisson family for which the LHS cannot be lower than 0, or the logit
family for which the LHS lies within 0 and 1.
In that case the fixed-effects (FEs) with only-0 LHS would lead to infinite coefficients
(FE = -Inf would explain perfectly the LHS).
The value \code{fixef.rm="infinite_coef"} removes all observations associated to FEs with
infinite coefficients.

If "perfect_fit", it is equivalent to "singletons" and "infinite_coef" combined.
That means all observations that are perfectly explained by the FEs are removed.

If "none": no observation is removed.

Note that whatever the value of this options: the coefficient estimates
will remain the same. It only affects inference (the standard-errors).

The algorithm is recursive, meaning that, e.g. in the presence of several fixed-effects (FEs),
removing singletons in one FE can create singletons (or perfect fits) in another FE.
The algorithm continues until there is no singleton/perfect-fit remaining.}

\item{fixef.tol}{Precision used to obtain the fixed-effects. Defaults to \code{1e-5}.
It corresponds to the maximum absolute difference allowed between two coefficients
of successive iterations. Argument \code{fixef.tol} cannot be lower
than \code{10000*.Machine$double.eps}. Note that this parameter is dynamically
controlled by the algorithm.}

\item{fixef.iter}{Maximum number of iterations in fixed-effects algorithm
(only in use for 2+ fixed-effects). Default is 10000.}

\item{collin.tol}{Numeric scalar, default is \code{1e-9}. Threshold deciding when variables should
be considered collinear and subsequently removed from the estimation. Higher values means more
variables will be removed (if there is presence of collinearity). One signal of presence of
collinearity is t-stats that are extremely low (for instance when t-stats < 1e-3).}

\item{lean}{Logical scalar, default is \code{FALSE}. If \code{TRUE} then all large objects are removed
from the returned result: this will save memory but will block the possibility to
use many methods. It is recommended to use the arguments \code{se} or \code{cluster} to
obtain the appropriate standard-errors at estimation time, since obtaining different
SEs won't be possible afterwards.}

\item{verbose}{Integer. Higher values give more information. In particular,
it can detail the number of iterations in the demeaning algorithm
(the first number is the left-hand-side, the other numbers are the right-hand-side variables).}

\item{warn}{Logical, default is \code{TRUE}. Whether warnings should be displayed
(concerns warnings relating to convergence state).}

\item{fixef.keep_names}{Logical or \code{NULL} (default). When you combine different
variables to transform them into a single fixed-effects you can do
e.g. \code{y ~ x | paste(var1, var2)}.
The algorithm provides a shorthand to do the same operation: \code{y ~ x | var1^var2}.
Because pasting variables is a costly operation, the internal algorithm may use a
numerical trick to hasten the process. The cost of doing so is that you lose the labels.
If you are interested in getting the value of the fixed-effects coefficients
after the estimation, you should use \code{fixef.keep_names = TRUE}. By default it is
equal to \code{TRUE} if the number of unique values is lower than 50,000, and to \code{FALSE}
otherwise.}

\item{demeaned}{Logical, default is \code{FALSE}. Only used in the presence of fixed-effects: should
the centered variables be returned? If \code{TRUE}, it creates the items
\code{y_demeaned} and \code{X_demeaned}.}

\item{mem.clean}{Logical scalar, default is \code{FALSE}. Only to be used if the data set is
large compared to the available RAM. If \code{TRUE} then intermediary objects are removed as
much as possible and \code{\link{gc}} is run before each substantial C++ section in the internal
code to avoid memory issues.}

\item{glm.iter}{Number of iterations of the glm algorithm. Default is 25.}

\item{glm.tol}{Tolerance level for the glm algorithm. Default is \code{1e-8}.}

\item{demeaner}{\code{NULL} (default) or a demeaning backend object created by
\code{\link{MapDemeaner}} or \code{\link{LsmrDemeaner}}. If \code{NULL}, the native MAP backend is
used with \code{fixef.tol} and \code{fixef.iter}.}

\item{data.save}{Logical scalar, default is \code{FALSE}. If \code{TRUE}, the data used for
the estimation is saved within the returned object. Hence later calls to predict(),
vcov(), etc..., will be consistent even if the original data has been modified
in the meantime.
This is especially useful for estimations within loops, where the data changes
at each iteration, such that postprocessing can be done outside the loop without issue.}

\item{reset}{Logical scalar, default is \code{FALSE}. Whether to reset all values.}
}
\value{
The function \code{getFixest_estimation} returns the currently set global defaults.
}
\description{
This function sets globally the default arguments of fixest estimations.
}
\examples{

#
# Example: removing singletons is FALSE by default
#

# => changing this default

# Let's create data with singletons
base = iris
names(base) = c("y", "x1", "x2", "x3", "species")
base$fe_singletons = as.character(base$species)
base$fe_singletons[1:5] = letters[1:5]

res          = feols(y ~ x1 + x2 | fe_singletons, base)
res_noSingle = feols(y ~ x1 + x2 | fe_singletons, base, fixef.rm = "single")

# New defaults
setFixest_estimation(fixef.rm = "single")
res_newDefault = feols(y ~ x1 + x2 | fe_singletons, base)

etable(res, res_noSingle, res_newDefault)

# Resetting the defaults
setFixest_estimation(reset = TRUE)



}
