% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCOV.R
\name{setFixest_ssc}
\alias{setFixest_ssc}
\title{Sets the default values for the small sample correction}
\usage{
setFixest_ssc(ssc_type = NULL, vcov_names = "iid")
}
\arguments{
\item{ssc_type}{Either \code{NULL} (default), or an object of class \code{ssc_type} obtained
with the function \code{\link{ssc}}, or old options values (of class \code{setFixest_ssc}).
By default, if \code{NULL}, it resets the SSC to their default value for the selected VCOVs.
Otherwise it sets the default SSC for the selected VCOVs to the value returned by
the function ssc().}

\item{vcov_names}{Character vector corresponding to the keywords of the VCOVs for which
to change the default SSC. By default it is equal to \code{"iid"}. Some common VCOV names are:
"iid", "hetero", "cluster", "twoway", "newey", "driscoll", "conley".}
}
\value{
This functions invisibly returns the list of old settings, a list of class \code{setFixest_ssc}.
}
\description{
Sets the default values for the Small Sample Correction (SSC) for specific \code{fixest} VCOVs
}
\examples{

# Estimation with current default values for the small sample correction
feols(Sepal.Length ~ Petal.Length + Petal.Width, iris)

# looking at the default SSC
getFixest_ssc("iid")

# 1) setting new default (and saving the previous opts)
old_opts = setFixest_ssc(ssc(K.adj = FALSE), "iid")
# => the SEs/t-stat differ
feols(Sepal.Length ~ Petal.Length + Petal.Width, iris)

getFixest_ssc("iid")

# 2) resetting to the old values
setFixest_ssc(old_opts)
getFixest_ssc("iid")


}
\seealso{
\code{\link{vcov.fixest}}, \code{\link{ssc}}, \code{\link{getFixest_ssc}}
}
