% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{sigma.fixest}
\alias{sigma.fixest}
\title{Residual standard deviation of \code{fixest} estimations}
\usage{
\method{sigma}{fixest}(object, ...)
}
\arguments{
\item{object}{A \code{fixest} object.}

\item{...}{Not currently used.}
}
\value{
Returns a numeric scalar.
}
\description{
Extract the estimated standard deviation of the errors from \code{fixest} estimations.
}
\examples{

est = feols(Petal.Length ~ Petal.Width, iris)
sigma(est)



}
\seealso{
\code{\link[fixest]{feols}}, \code{\link[fixest:feglm]{fepois}}, \code{\link[fixest]{feglm}}, \code{\link[fixest:femlm]{fenegbin}}, \code{\link[fixest]{feNmlm}}.
}
