% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sigma.fixest2}
\alias{sigma.fixest2}
\title{Residual standard deviation of \code{fixest2} estimations}
\usage{
\method{sigma}{fixest2}(object, ...)
}
\arguments{
\item{object}{A \code{fixest2} object.}

\item{...}{Not currently used.}
}
\value{
Returns a numeric scalar.
}
\description{
Extract the estimated standard deviation of the errors from \code{fixest2} estimations.
}
\examples{

est <- feols(Petal.Length ~ Petal.Width, iris)
sigma(est)

}
\seealso{
\code{\link{feols}}, \code{\link[=feglm]{fepois}}, \code{\link{feglm}}, \code{\link[=femlm]{fenegbin}}, \code{\link{feNmlm}}.
}
