% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{[[.fixest2_multi}
\alias{[[.fixest2_multi}
\title{Extracts one element from a \code{fixest2_multi} object}
\usage{
\method{[[}{fixest2_multi}(x, i)
}
\arguments{
\item{x}{A \code{fixest2_multi} object, obtained from a \code{fixest2} estimation leading to
multiple results.}

\item{i}{An integer scalar. The identifier of the estimation to extract.}
}
\value{
A \code{fixest2} object is returned.
}
\description{
Extracts single elements from multiple \code{fixest2} estimations.
}
\examples{

base <- iris
names(base) <- c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res <- feols(y ~ csw(x1, x2, x3), base, split = ~species)

# The first estimation
res[[1]]

# The second one, etc
res[[2]]
}
\seealso{
The main fixest2 estimation functions: \code{\link{feols}}, \code{\link[=feglm]{fepois}},
\code{\link[=femlm]{fenegbin}}, \code{\link{feglm}}, \code{\link{feNmlm}}. Tools for mutliple fixest2
estimations: \code{\link{summary.fixest2_multi}}, \code{\link{print.fixest2_multi}}, \code{\link{as.list.fixest2_multi}},
\code{\link[fixest2]{sub-sub-.fixest2_multi}}, \code{\link[fixest2]{sub-.fixest2_multi}}.
}
