% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated_funs.R
\name{summary.fixest.obs2remove}
\alias{summary.fixest.obs2remove}
\title{Summary method for fixest.obs2remove objects}
\usage{
\method{summary}{fixest.obs2remove}(object, ...)
}
\arguments{
\item{object}{A \code{fixest.obs2remove} object obtained from function \code{\link[fixest]{obs2remove}}.}

\item{...}{Not currently used.}
}
\description{
This function synthesizes the information of function \code{\link[fixest]{obs2remove}}. It reports the number of observations to be removed as well as the number of fixed-effects removed per fixed-effect dimension.
}
\examples{
base = iris
# v6: Petal.Length with only 0 values for 'setosa'
base$v6 = base$Petal.Length
base$v6[base$Species == "setosa"] = 0

x = obs2remove(v6 ~ Species, base)
summary(x)

}
\seealso{
\code{\link{fixest-deprecated}}
}
\keyword{internal}
