% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{summary.fixest_multi}
\alias{summary.fixest_multi}
\title{Summary for fixest_multi objects}
\usage{
\method{summary}{fixest_multi}(object, type = "short", ...)
}
\arguments{
\item{object}{A \code{fixest_multi} object, obtained from a \code{fixest} estimation leading to multiple results.}

\item{type}{A character either equal to \code{"short"}, \code{"long"}, \code{"compact"}, or \code{"se_compact"}. If \code{short}, only the table of coefficients is displayed for each estimation. If \code{long}, then the full results are displayed for each estimation. If \code{compact}, a \code{data.frame} is returned with one line per model and the formatted coefficients + standard-errors in the columns. If \code{se_compact}, a \code{data.frame} is returned with one line per model, one numeric column for each coefficient and one numeric column for each standard-error.}

\item{...}{Other arguments to be passed to \code{\link[fixest]{summary.fixest}}.}
}
\value{
It returns either an object of class \code{fixest_multi} (if \code{type} equals \code{short} or \code{long}), either a \code{data.frame} (if type equals \code{compact} or \code{se_compact}).
}
\description{
Summary information for fixest_multi objects. In particular, this is used to specify the type of standard-errors to be computed.
}
\examples{

base = iris
names(base) = c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res = feols(y ~ csw(x1, x2, x3), base, split = ~species)

# By default, the type is "short"
# You can stil use the arguments from summary.fixest
summary(res, cluster = ~ species)

summary(res, type = "long")

summary(res, type = "compact")

summary(res, type = "se_compact")


}
\seealso{
The main fixest estimation functions: \code{\link[fixest]{feols}}, \code{\link[fixest:feglm]{fepois}}, \code{\link[fixest:femlm]{fenegbin}}, \code{\link[fixest]{feglm}}, \code{\link[fixest]{feNmlm}}. Tools for mutliple fixest estimations: \code{\link[fixest]{summary.fixest_mutli}}, \code{\link[fixest]{print.fixest_mutli}}, \code{\link[fixest]{as.list.fixest_multi}}, \code{\link[fixest]{sub-sub-.fixest_multi}}, \code{\link[fixest]{sub-.fixest_multi}}, \code{\link[fixest]{cash-.fixest_multi}}.
}
