% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{terms.fixest2}
\alias{terms.fixest2}
\title{Extract the terms}
\usage{
\method{terms}{fixest2}(x, ...)
}
\arguments{
\item{x}{A \code{fixest2} object. Obtained using the functions \code{\link{femlm}}, \code{\link{feols}} or \code{\link{feglm}}.}

\item{...}{Not currently used.}
}
\value{
An object of class \code{c("terms", "formula")} which contains the terms representation of a
symbolic model.
}
\description{
This function extracts the terms of a \code{fixest2} estimation, excluding the fixed-effects part.
}
\examples{

# simple estimation on iris data, using "Species" fixed-effects
res <- feols(Sepal.Length ~ Sepal.Width * Petal.Length +
  Petal.Width | Species, iris)

# Terms of the linear part
terms(res)

}
