% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\docType{data}
\name{trade}
\alias{trade}
\title{Trade data sample}
\format{
\code{trade} is a data frame with 38,325 observations and 6 variables named \code{Destination}, \code{Origin}, \code{Product}, \code{Year}, \code{dist_km} and \code{Euros}.

\itemize{
\item{Origin: 2-digits codes of the countries of origin of the trade flow.}
\item{Destination: 2-digits codes of the countries of destination of the trade flow.}
\item{Products: Number representing the product categories (from 1 to 20).}
\item{Year: Years from 2007 to 2016}
\item{dist_km: Geographic distance in km between the centers of the countries of origin and destination.}
\item{Euros: The total amount in euros of the trade flow for the specific year/product category/origin-destination country pair.}

}
}
\source{
This data has been extrated from Eurostat on October 2017.
}
\usage{
data(trade)
}
\description{
This data reports trade information between countries of the European Union (EU15).
}
\keyword{datasets}
