% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{weights.fixest2}
\alias{weights.fixest2}
\title{Extracts the weights from a \code{fixest2} object}
\usage{
\method{weights}{fixest2}(object, ...)
}
\arguments{
\item{object}{A \code{fixest2} object.}

\item{...}{Not currently used.}
}
\value{
Returns a vector of the same length as the number of observations in the original data set.
Ignored observations due to NA or perfect fit are re-introduced and their weights set to NA.
}
\description{
Simply extracts the weights used to estimate a \code{fixest2} model.
}
\examples{

est <- feols(Petal.Length ~ Petal.Width, iris, weights = ~ as.integer(Sepal.Length) - 3.99)
weights(est)

}
\seealso{
\code{\link{feols}}, \code{\link[=feglm]{fepois}}, \code{\link{feglm}}, \code{\link[=femlm]{fenegbin}}, \code{\link{feNmlm}}.
}
