% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{as.list.fixest_multi}
\alias{as.list.fixest_multi}
\title{Transforms a fixest_multi object into a list}
\usage{
\method{as.list}{fixest_multi}(x, ...)
}
\arguments{
\item{x}{A `fixest_multi` object, obtained from a `fixest` estimation leading to multiple results.}

\item{...}{Not currently used.}
}
\value{
Returns a list containing all the results of the multiple estimations.
}
\description{
Extracts the results from a `fixest_multi` object and place them into a list.
}
\examples{

base <- iris
names(base) <- c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res <- feols(y ~ csw(x1, x2, x3), base, split = ~species)

# All the results at once
as.list(res)

}
\seealso{
The main fixest estimation functions: [`feols`], [`fepois`][fixest2::feglm], [`fenegbin`][fixest2::femlm], [`feglm`], [`feNmlm`]. Tools for mutliple fixest estimations: [`summary.fixest_multi`], [`print.fixest_multi`], [`as.list.fixest_multi`], \code{\link[fixest]{sub-sub-.fixest_multi}}, \code{\link[fixest]{sub-.fixest_multi}}.
}
