% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.fixest}
\alias{coef.fixest}
\alias{coefficients.fixest}
\title{Extracts the coefficients from a `fixest` estimation}
\usage{
\method{coef}{fixest}(object, keep, drop, order, collin = FALSE, agg = TRUE, ...)

\method{coefficients}{fixest}(object, keep, drop, order, collin = FALSE, agg = TRUE, ...)
}
\arguments{
\item{object}{A `fixest` object. Obtained using the functions [`femlm`], [`feols`] or [`feglm`].}

\item{collin}{Logical, default is `FALSE`. Whether the coefficients removed because of collinearity should be also returned as `NA`. It cannot be used when coefficients aggregation is also used.}

\item{agg}{Logical scalar, default is `TRUE`. If the coefficients of the estimation have been aggregated, whether to report the aggregated coefficients. If `FALSE`, the raw coefficients will be returned.}

\item{...}{Not currently used.}
}
\value{
This function returns a named numeric vector.
}
\description{
This function extracts the coefficients obtained from a model estimated with [`femlm`], [`feols`] or [`feglm`].
}
\details{
The coefficients are the ones that have been found to maximize the log-likelihood of the specified model. More information can be found on the models from the estimations help pages: [`femlm`], [`feols`] or [`feglm`].

Note that if the model has been estimated with fixed-effects, to obtain the fixed-effect coefficients, you need to use the function [`fixef.fixest`].
}
\examples{

# simple estimation on iris data, using "Species" fixed-effects
res <- femlm(Sepal.Length ~ Sepal.Width + Petal.Length +
    Petal.Width | Species, iris)

# the coefficients of the variables:
coef(res)

# the fixed-effects coefficients:
fixef(res)

}
\seealso{
See also the main estimation functions [`femlm`], [`feols`] or [`feglm`]. [`summary.fixest`], [`confint.fixest`], [`vcov.fixest`], [`fixef.fixest`].
}
\author{
Laurent Berge
}
