% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{coef.fixest_multi}
\alias{coef.fixest_multi}
\alias{coefficients.fixest_multi}
\title{Extracts the coefficients of fixest_multi objects}
\usage{
\method{coef}{fixest_multi}(
  object,
  keep,
  drop,
  order,
  collin = FALSE,
  long = FALSE,
  na.rm = TRUE,
  ...
)

\method{coefficients}{fixest_multi}(
  object,
  keep,
  drop,
  order,
  collin = FALSE,
  long = FALSE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{object}{A `fixest_multi` object. Obtained from a multiple estimation.}

\item{collin}{Logical, default is `FALSE`. Whether the coefficients removed because of collinearity should be also returned as `NA`. It cannot be used when coefficients aggregation is also used.}

\item{long}{Logical, default is `FALSE`. Whether the results should be displayed in a long format.}

\item{na.rm}{Logical, default is `TRUE`. Only applies when `long = TRUE`: whether to remove the coefficients with `NA` values.}

\item{...}{Not currently used.}
}
\description{
Utility to extract the coefficients of multiple estimations and rearrange them into a matrix.
}
\examples{

base <- iris
names(base) <- c("y", "x1", "x2", "x3", "species")

# A multiple estimation
est <- feols(y ~ x1 + csw0(x2, x3), base)

# Getting all the coefficients at once,
# each row is a model
coef(est)

# Example of keep/drop/order
coef(est, keep = "Int|x1", order = "x1")


# To change the order of the model, use fixest_multi
# extraction tools:
coef(est[rhs = .N:1])

# collin + long + na.rm
base$x1_bis <- base$x1 # => collinear
est <- feols(y ~ x1_bis + csw0(x1, x2, x3), base, split = ~species)

# does not display x1 since it is always collinear
coef(est)
# now it does
coef(est, collin = TRUE)

# long
coef(est, long = TRUE)

# long but balanced (with NAs then)
coef(est, long = TRUE, na.rm = FALSE)

}
