% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coeftable.default}
\alias{coeftable.default}
\alias{se.default}
\alias{tstat.default}
\alias{pvalue.default}
\alias{se.matrix}
\title{Extracts the coefficients table from an estimation}
\usage{
\method{coeftable}{default}(object, keep, drop, order, ...)

\method{se}{default}(object, keep, drop, order, ...)

\method{tstat}{default}(object, keep, drop, order, ...)

\method{pvalue}{default}(object, keep, drop, order, ...)

\method{se}{matrix}(object, keep, drop, order, ...)
}
\arguments{
\item{object}{The result of an estimation (a fitted model object). Note that this function is made to work with `fixest` objects so it may not work for the specific model you provide.}

\item{...}{Other arguments that will be passed to `summary`.

First the method summary is applied if needed, then the coefficients table is extracted from its output.

The default method is very naive and hopes that the resulting coefficients table contained in the summary of the fitted model is well formed: this assumption is very often wrong. Anyway, there is no development intended since the coeftable/se/pvalue/tstat series of methods is only intended to work well with `fixest` objects. To extract the coefficients table from fitted models in a general way, it's better to use [tidy from broom](https://broom.tidymodels.org/).}
}
\value{
Returns a matrix (`coeftable`) or vectors.
}
\description{
Default method to extracts the coefficients table and its sub-components from an estimation.
}
\section{Functions}{
\itemize{
\item \code{se(default)}: Extracts the standard-errors from an estimation

\item \code{tstat(default)}: Extracts the standard-errors from an estimation

\item \code{pvalue(default)}: Extracts the p-values from an estimation

\item \code{se(matrix)}: Extracts the standard-errors from a VCOV matrix

}}
\examples{

# NOTA: This function is really made to handle fixest objects
# The default methods works for simple structures, but you'd be
# likely better off with broom::tidy for other models

est <- lm(mpg ~ cyl, mtcars)
coeftable(est)

se(est)

}
