% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{coeftable.fixest_multi}
\alias{coeftable.fixest_multi}
\alias{se.fixest_multi}
\alias{tstat.fixest_multi}
\alias{pvalue.fixest_multi}
\title{Extracts the coefficients tables from `fixest_multi` estimations}
\usage{
\method{coeftable}{fixest_multi}(
  object,
  vcov = NULL,
  keep = NULL,
  drop = NULL,
  order = NULL,
  long = FALSE,
  wide = FALSE,
  ...
)

\method{se}{fixest_multi}(
  object,
  vcov = NULL,
  keep = NULL,
  drop = NULL,
  order = NULL,
  long = FALSE,
  ...
)

\method{tstat}{fixest_multi}(
  object,
  vcov = NULL,
  keep = NULL,
  drop = NULL,
  order = NULL,
  long = FALSE,
  ...
)

\method{pvalue}{fixest_multi}(
  object,
  vcov = NULL,
  keep = NULL,
  drop = NULL,
  order = NULL,
  long = FALSE,
  ...
)
}
\arguments{
\item{object}{A `fixest_multi` object, coming from a `fixest` multiple estimation.}

\item{long}{Logical scalar, default is `FALSE`. If `TRUE`, then all the information is stacked, with two columns containing the information: `"param"` and `"value"`. The column `param` contains the values `coef`/`se`/`tstat`/`pvalue`.}

\item{wide}{A logical scalar, default is `FALSE`. If `TRUE`, then a list is returned: the elements of the list are coef/se/tstat/pvalue. Each element of the list is a wide table with a column per coefficient.}

\item{...}{Other arguments to be passed to [`summary.fixest`].}
}
\value{
It returns a `data.frame` containing the coefficients tables (or just the se/pvalue/tstat) along with the information on which model was estimated.

If `wide = TRUE`, then a list is returned. The elements of the list are coef/se/tstat/pvalue. Each element of the list is a wide table with a column per coefficient.

If `long = TRUE`, then all the information is stacked. This removes the 4 columns containing the coefficient estimates to the p-values, and replace them with two new columns: `"param"` and `"value"`. The column `param` contains the values `coef`/`se`/`tstat`/`pvalue`, and the column `values` the associated numerical information.
}
\description{
Series of methods to extract the coefficients table or its sub-components from a `fixest_multi` objects (i.e. the outcome of multiple estimations).
}
\section{Functions}{
\itemize{
\item \code{se(fixest_multi)}: Extracts the standard-errors from `fixest_multi` estimations

\item \code{tstat(fixest_multi)}: Extracts the t-stats from `fixest_multi` estimations

\item \code{pvalue(fixest_multi)}: Extracts the p-values from `fixest_multi` estimations

}}
\examples{

base <- setNames(iris, c("y", "x1", "x2", "x3", "species"))
est_multi <- feols(y ~ csw(x.[, 1:3]), base, split = ~species)

# we get all the coefficient tables at once
coeftable(est_multi)

# Now just the standard-errors
se(est_multi)

# wide = TRUE => leads toa  list of wide tables
coeftable(est_multi, wide = TRUE)

# long = TRUE, all the information is stacked
coeftable(est_multi, long = TRUE)

}
