% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\name{coeftable}
\alias{coeftable}
\alias{se}
\alias{pvalue}
\alias{tstat}
\alias{pvalue.fixest}
\alias{tstat.fixest}
\alias{se.fixest}
\title{Extracts the coefficients table from an estimation}
\usage{
coeftable(object, ...)

se(object, ...)

pvalue(object, ...)

tstat(object, ...)

\method{pvalue}{fixest}(
  object,
  vcov = NULL,
  ssc = NULL,
  cluster = NULL,
  keep = NULL,
  drop = NULL,
  order = NULL,
  ...
)

\method{tstat}{fixest}(
  object,
  vcov = NULL,
  ssc = NULL,
  cluster = NULL,
  keep = NULL,
  drop = NULL,
  order = NULL,
  ...
)

\method{se}{fixest}(
  object,
  vcov = NULL,
  ssc = NULL,
  cluster = NULL,
  keep = NULL,
  drop = NULL,
  order = NULL,
  ...
)
}
\arguments{
\item{object}{An estimation (fitted model object), e.g. a \code{fixest} object.}

\item{...}{Other arguments to the methods.}
}
\value{
Returns a matrix (\code{coeftable}) or vectors.
}
\description{
Methods to extracts the coefficients table and its sub-components from an estimation.
}
\section{Methods (by class)}{
\itemize{
\item \code{fixest}: Extracts the p-value of an estimation

\item \code{fixest}: Extracts the t-statistics of an estimation

\item \code{fixest}: Extracts the standard-error of an estimation
}}

\examples{

est = lm(mpg ~ cyl, mtcars)
coeftable(est)

}
\seealso{
Please look at the \code{\link[fixest]{coeftable.fixest}} page for more detailed information.
}
