% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\name{fixef_reexported}
\alias{fixef_reexported}
\alias{fixef}
\title{Functions exported from \pkg{nlme} to implement \pkg{fixest} methods}
\description{
The package \pkg{fixest} uses the \code{fixef} method from \pkg{nlme}. Unfortunately, re-exporting this method is required in order not to attach package \pkg{nlme}.
}
\details{
\itemize{
\item Here is the help from package \pkg{nlme}: \code{\link[nlme:fixed.effects]{fixef}}. The help from package \pkg{fixest} is here: \code{\link[fixest]{fixef.fixest}}.
}
}
\note{
I could find this workaround thanks to the package \pkg{plm}.
}
\keyword{internal}
