% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\name{fixest_startup_msg}
\alias{fixest_startup_msg}
\title{Permanently removes the fixest package startup message}
\usage{
fixest_startup_msg(x)
}
\arguments{
\item{x}{Logical, no default. If \code{FALSE}, the package startup message is removed.}
}
\description{
Package startup messages can be very annoying, although sometimes they can be necessary. Use this function to prevent \code{fixest}'s package startup message from popping when loading. This will be specific to your current project.
}
\details{
Note that this function is introduced to cope with the first \code{fixest} startup message (in version 0.9.0). In the future, all startup messages may be removed, but the function will still exist.

This function works by adding a variable in the \code{.Renviron} file, so it is very lightweight and project-specific.
}
