% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{formula.fixest}
\alias{formula.fixest}
\title{Extract the formula of a `fixest` fit}
\usage{
\method{formula}{fixest}(x, type = c("full", "linear", "iv", "NL"), ...)
}
\arguments{
\item{x}{An object of class `fixest`. Typically the result of a [`femlm`], [`feols`] or [`feglm`] estimation.}

\item{type}{A character scalar. Default is `type = "full"` which gives back a formula containing the linear part of the model along with the fixed-effects (if any) and the IV part (if any). If `type = "linear"` then only the linear formula is returned. If `type = "NL"` then only the non linear in parameters part is returned.}

\item{...}{Not currently used.}
}
\value{
It returns a formula.
}
\description{
This function extracts the formula from a `fixest` estimation (obtained with [`femlm`], [`feols`] or [`feglm`]). If the estimation was done with fixed-effects, they are added in the formula after a pipe (\dQuote{|}). If the estimation was done with a non linear in parameters part, then this will be added in the formula in between `I()`.
}
\examples{

# simple estimation on iris data, using "Species" fixed-effects
res <- femlm(Sepal.Length ~ Sepal.Width + Petal.Length +
    Petal.Width | Species, iris)

# formula with the fixed-effect variable
formula(res)

# linear part without the fixed-effects
formula(res, "linear")

}
\seealso{
See also the main estimation functions [`femlm`], [`feols`] or [`feglm`]. [`model.matrix.fixest`], [`update.fixest`], [`summary.fixest`], [`vcov.fixest`].
}
\author{
Laurent Berge
}
