% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{hatvalues.fixest}
\alias{hatvalues.fixest}
\title{Hat values for \code{fixest} objects}
\usage{
\method{hatvalues}{fixest}(model, exact = TRUE, p = 1000, ...)
}
\arguments{
\item{model}{A fixest object. For instance from feols or feglm.}

\item{exact}{Logical scalar, default is \code{TRUE}. Whether the diagonals of the projection matrix should be calculated exactly. If \code{FALSE}, then it will be}

\item{p}{Numeric scala, default is 1000. This is only used when \code{exact == FALSE}. This determined the number of bootstrap samples used to estimate the projection matrix.}

\item{...}{Not currently used.}
}
\value{
Returns a vector of the same length as the number of observations used in the estimation.
}
\description{
Computes the hat values for \code{\link{feols}} or \code{\link{feglm}} estimations.
}
\details{
Hat values are not available for \code{\link[=femlm]{fenegbin}}, \code{\link{femlm}} and \code{\link{feNmlm}} estimations.

When \code{exact == FALSE}, the Johnson-Lindenstrauss approximation (LA) algorithm is used which approximates the diagonals of the projection matrix. For more precision (but longer time), increase the value of \code{p}.
}
\examples{

est = feols(Petal.Length ~ Petal.Width + Sepal.Width, iris)
head(hatvalues(est))


}
