% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\name{n_unik}
\alias{n_unik}
\alias{print.vec_n_unik}
\title{Prints the number of unique elements}
\usage{
n_unik(x, fml)

\method{print}{vec_n_unik}(x, ...)
}
\arguments{
\item{x}{A data.frame or a vector.}

\item{fml}{A one-sided formula containing the variables for which to count the number of unique values. You can use \code{.N} to get the number of observations. You can combine variables with the "^" operator. For example: \code{fml = ~ .N + id + id^period} will report the number of observations, the number of unique values of \code{id} and the number of unique \code{id} x \code{period} pairs. You can sub select variables with \code{[]}, like in \code{id[!is.na(period)]}. You can use the special functions \code{sw} and \code{sw0}, like in \code{~id^sw0(period) + period[sw0(is.na(id))]} which would lead to \code{~id + id^period + period + period[is.na(id)]}.}
}
\value{
It returns a vector containing the number of unique values per element.
}
\description{
This utility tool displays the number of unique elements in a data.frame as well as their number of NA values.
}
\examples{

data = base_did
data$x1.L1 = round(lag(x1~id+period, 1, data))

# By default, just the formatted number of observations
n_unik(data)

# Or the nber of unique elements of a vector
n_unik(data$id)

# number of unique id values and id x period pairs
n_unik(data, ~.N + id + id^period)

# using sub selection + sw0
n_unik(data, ~.N + period[sw0(!is.na(x1.L1))])


}
