% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{print.fixest_multi}
\alias{print.fixest_multi}
\title{Print method for fixest_multi objects}
\usage{
\method{print}{fixest_multi}(x, ...)
}
\arguments{
\item{x}{A \code{fixest_multi} object, obtained from a \code{fixest} estimation leading to multiple results.}

\item{...}{Other arguments to be passed to \code{\link[fixest2]{summary.fixest_multi}}.}
}
\description{
Displays summary information on fixest_multi objects in the R console.
}
\examples{

base <- iris
names(base) <- c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res <- feols(y ~ csw(x1, x2, x3), base, split = ~species)

# Let's print all that
res

}
\seealso{
The main fixest estimation functions: \code{\link[fixest2]{feols}},
\code{\link[fixest2:feglm]{fepois}}, \code{\link[fixest2:femlm]{fenegbin}},
\code{\link[fixest2]{feglm}}, \code{\link[fixest2]{feNmlm}}. Tools for multiple
fixest estimations: \code{\link[fixest2]{summary.fixest_multi}},
\code{\link[fixest2]{print.fixest_multi}},
\code{\link[fixest2]{as.list.fixest_multi}},
\code{\link[fixest2]{sub-sub-.fixest_multi}},
\code{\link[fixest2]{sub-.fixest_multi}},
\code{\link[fixest2]{cash-.fixest_multi}}.
}
