% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{sandwich_reexported}
\alias{sandwich_reexported}
\alias{estfun}
\alias{bread}
\title{Functions exported from \pkg{sandwich} to implement \pkg{fixest} methods}
\description{
The package \pkg{fixest} does not use \code{estfun} or \code{bread} from \pkg{sandwich}, but these methods have been implemented to allow users to leverage the variances from \pkg{sandwich}.
}
\details{
\itemize{
\item Here is the help from package \pkg{sandwich}: \code{\link[sandwich:estfun]{estfun}} and \code{\link[sandwich:bread]{bread}}. The help from package \pkg{fixest} is here: \code{\link[fixest]{estfun.fixest}} and \code{\link[fixest]{bread.fixest}}.
}
}
\keyword{internal}
