% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\docType{data}
\name{setFixest_notes}
\alias{setFixest_notes}
\alias{getFixest_notes}
\title{Sets/gets whether to display notes in \code{fixest} estimation functions}
\format{
An object of class \code{function} of length 1.
}
\usage{
setFixest_notes(x)

getFixest_notes
}
\arguments{
\item{x}{A logical. If \code{FALSE}, then notes are permanently removed.}
}
\description{
Sets/gets the default values of whether notes (informing for NA and observations removed) should be displayed in \code{fixest} estimation functions.
}
\examples{

# Change default with
setFixest_notes(FALSE)

# Back to default which is TRUE
getFixest_notes()

}
\author{
Laurent Berge
}
\keyword{datasets}
