% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\name{setFixest_nthreads}
\alias{setFixest_nthreads}
\alias{getFixest_nthreads}
\title{Sets/gets the number of threads to use in \code{fixest} functions}
\usage{
setFixest_nthreads(nthreads)

getFixest_nthreads()
}
\arguments{
\item{nthreads}{An integer strictly greater than one and lower than the maximum number of threads (if OpenMP is available). If missing, the default is the maximum number of threads minus two.}
}
\description{
Sets/gets the default number of threads to used in \code{fixest} estimation functions. The default is the maximum number of threads minus two.
}
\examples{

\donttest{
# Gets the current number of threads
getFixest_nthreads()
# To set multi-threading off:
setFixest_nthreads(1)
# To set it back to default:
setFixest_nthreads()
}


}
\author{
Laurent Berge
}
