% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_model_matrix.R
\name{sparse_model_matrix}
\alias{sparse_model_matrix}
\title{Design matrix of a \code{fixest} object returned in sparse format}
\usage{
sparse_model_matrix(
  object,
  data,
  type = "rhs",
  na.rm = TRUE,
  collin.rm = TRUE,
  combine = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{fixest} object. Obtained using the functions \code{\link{femlm}}, \code{\link{feols}} or \code{\link{feglm}}.}

\item{data}{If missing (default) then the original data is obtained by evaluating the \code{call}. Otherwise, it should be a \code{data.frame}.}

\item{type}{Character vector or one sided formula, default is "rhs". Contains the type of matrix/data.frame to be returned. Possible values are: "lhs", "rhs", "fixef", "iv.rhs1" (1st stage RHS), "iv.rhs2" (2nd stage RHS), "iv.endo" (endogenous vars.), "iv.exo" (exogenous vars), "iv.inst" (instruments).}

\item{na.rm}{Default is \code{TRUE}. Should observations with NAs be removed from the matrix?}

\item{collin.rm}{Logical scalar, default is \code{TRUE}. Whether to remove variables that were found to be collinear during the estimation. Beware: it does not perform a collinearity check.}

\item{combine}{Logical scalar, default is \code{TRUE}. Whether to combine each resulting sparse matrix}

\item{...}{Not currently used.}
}
\value{
It returns either a single sparse matrix a list of matrices, depending whether \code{combine} is \code{TRUE} or \code{FALSE}. The sparse matrix is of class \code{dgCMatrix} from the \code{Matrix} package.
}
\description{
This function creates the left-hand-side or the right-hand-side(s) of a \code{\link{femlm}}, \code{\link{feols}} or \code{\link{feglm}} estimation. Note that this function currently does not accept a formula
}
\examples{

est = feols(wt ~ i(vs) + hp | cyl, mtcars)
sparse_model_matrix(est)


}
\seealso{
See also the main estimation functions \code{\link{femlm}}, \code{\link{feols}} or \code{\link{feglm}}. \code{\link{formula.fixest}}, \code{\link{update.fixest}}, \code{\link{summary.fixest}}, \code{\link{vcov.fixest}}.
}
\author{
Laurent Berge, Kyle Butts
}
