% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_multi.R
\name{[[.fixest_multi}
\alias{[[.fixest_multi}
\title{Extracts one element from a `fixest_multi` object}
\usage{
\method{[[}{fixest_multi}(x, i)
}
\arguments{
\item{x}{A `fixest_multi` object, obtained from a `fixest` estimation leading to multiple results.}

\item{i}{An integer scalar. The identifier of the estimation to extract.}
}
\value{
A `fixest` object is returned.
}
\description{
Extracts single elements from multiple `fixest` estimations.
}
\examples{

base = iris
names(base) = c("y", "x1", "x2", "x3", "species")

# Multiple estimation
res = feols(y ~ csw(x1, x2, x3), base, split = ~species)

# The first estimation
res[[1]]

# The second one, etc
res[[2]]

}
\seealso{
The main fixest estimation functions: [`feols`], [`fepois`][fixest2::feglm], [`fenegbin`][fixest2::femlm], [`feglm`], [`feNmlm`]. Tools for mutliple fixest estimations: [`summary.fixest_multi`], [`print.fixest_multi`], [`as.list.fixest_multi`], \code{\link[fixest]{sub-sub-.fixest_multi}}, \code{\link[fixest]{sub-.fixest_multi}}.
}
