% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{terms.fixest}
\alias{terms.fixest}
\title{Extract the terms}
\usage{
\method{terms}{fixest}(x, ...)
}
\arguments{
\item{x}{A `fixest` object. Obtained using the functions [`femlm`], [`feols`] or [`feglm`].}

\item{...}{Not currently used.}
}
\value{
An object of class `c("terms", "formula")` which contains the terms representation of a symbolic model.
}
\description{
This function extracts the terms of a `fixest` estimation, excluding the fixed-effects part.
}
\examples{

# simple estimation on iris data, using "Species" fixed-effects
res = feols(Sepal.Length ~ Sepal.Width*Petal.Length +
            Petal.Width | Species, iris)

# Terms of the linear part
terms(res)


}
