% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{update.fixest}
\alias{update.fixest}
\title{Updates a `fixest` estimation}
\usage{
\method{update}{fixest}(object, fml.update, nframes = 1, evaluate = TRUE, ...)
}
\arguments{
\item{object}{A `fixest` object. Obtained using the functions [`femlm`], [`feols`] or [`feglm`].}

\item{fml.update}{Changes to be made to the original argument `fml`. See more information on [`update.formula`][stats::update.formula]. You can add/withdraw both variables and fixed-effects. E.g. `. ~ . + x2 | . + z2` would add the variable `x2` and the cluster `z2` to the former estimation.}

\item{nframes}{(Advanced users.) Defaults to 1. Number of frames up the stack where to perform the evaluation of the updated call. By default, this is the parent frame.}

\item{evaluate}{Logical, default is `TRUE`. If `FALSE`, only the updated call is returned.}

\item{...}{Other arguments to be passed to the functions [`femlm`], [`feols`] or [`feglm`].}
}
\value{
It returns a `fixest` object (see details in [`femlm`], [`feols`] or [`feglm`]).
}
\description{
Updates and re-estimates a `fixest` model (estimated with [`femlm`], [`feols`] or [`feglm`]). This function updates the formulas and use previous starting values to estimate a new `fixest` model. The data is obtained from the original `call`.
}
\examples{

# Example using trade data
data(trade)

# main estimation
est_pois = fepois(Euros ~ log(dist_km) | Origin + Destination, trade)

# we add the variable log(Year)
est_2 = update(est_pois, . ~ . + log(Year))

# we add another fixed-effect: "Product"
est_3 = update(est_2, . ~ . | . + Product)

# we remove the fixed-effect "Origin" and the variable log(dist_km)
est_4 = update(est_3, . ~ . - log(dist_km) | . - Origin)

# Quick look at the 4 estimations
etable(est_pois, est_2, est_3, est_4)

}
\seealso{
See also the main estimation functions [`femlm`], [`feols`] or [`feglm`]. [`predict.fixest`], [`summary.fixest`], [`vcov.fixest`], [`fixef.fixest`].
}
\author{
Laurent Berge
}
