% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCOV.R
\name{vcov_hetero}
\alias{vcov_hetero}
\title{Heteroskedasticity-Robust VCOV}
\usage{
vcov_hetero(x, type = "hc1", exact = TRUE, p = 500, ssc = NULL)
}
\arguments{
\item{x}{A \code{fixest} object.}

\item{type}{A string scalar. Either "HC1"/"HC2"/"HC3"}

\item{exact}{A logical scalar. Should the hat matrix be calculated exactly or approximated using the JLA algorithm? Default = TRUE. See \code{\link{hatvalues.fixest}} for details.}

\item{p}{A numeric scalar. The number of draws should be used when approximating the hat matrix? Note that larger \code{p} are more accurate, but slower. This is only used when \code{exact = TRUE}. Default is 500.}

\item{ssc}{An object returned by the function \code{\link{ssc}}. It specifies how to perform the small sample correction.}
}
\value{
If the first argument is a \code{fixest} object, then a VCOV is returned (i.e. a symmetric matrix).

If the first argument is not a \code{fixest} object, then a) implicitly the arguments are shifted to the left (i.e. \code{vcov_hetero("HC3")} is equivalent to \code{vcov_hetero(type = "HC3")} and b) a VCOV-\emph{request} is returned and NOT a VCOV. That VCOV-request can then be used in the argument \code{vcov} of various \code{fixest} functions (e.g. \code{\link{vcov.fixest}} or even in the estimation calls).
}
\description{
Computes the heteroskedasticity-robust VCOV of \code{fixest} objects.
}
\examples{

base = iris
names(base) = c("y", "x1", "x2", "x3", "species")
base$clu = rep(1:5, 30)

est = feols(y ~ x1 | species, base)

vcov_hetero(est, "hc1")
vcov_hetero(est, "hc2", ssc = ssc(adj = FALSE))
vcov_hetero(est, "hc3", ssc = ssc(adj = FALSE))

# Using approximate hatvalues
vcov_hetero(est, "hc3", exact = FALSE, p = 500)

}
\references{
Cameron AC, Gelbach JB, Miller DL (2011). "Robust Inference with Multiway Clustering." \emph{Journal of Business & Economic Statistics}, 29(2), 238-249. doi:10.1198/jbes.2010.07136.
}
\author{
Laurent Berge and Kyle Butts
}
